% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_fields.R
\name{add_missing_fields}
\alias{add_missing_fields}
\title{add_missing_fields}
\usage{
add_missing_fields(list_obj, template, reorder = FALSE)
}
\arguments{
\item{list_obj}{the list (for our purposes, probably your front-matter.yml after being read in)}

\item{template}{the list template to compare to}

\item{reorder}{do you want to reorder the resulting list, based on template order? default=F}
}
\value{
list_obj, with any missing fields added and reordered by template if requested
}
\description{
Internal function for \code{\link[=update_fm]{update_fm()}}. Adds missing list items to a list (list_obj), based on a template
}
