% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_compile.R
\name{batch_compile}
\alias{batch_compile}
\title{batch \code{\link[=compile_unit]{compile_unit()}}}
\usage{
batch_compile(
  shared_drive = "s",
  change_this = NULL,
  clean = FALSE,
  rebuild = FALSE,
  drive_reconnect = FALSE
)
}
\arguments{
\item{shared_drive}{passed to \code{\link[=pick_lesson]{pick_lesson()}} which shared drive do you want to find the lessons in? default= "s" Options:
\itemize{
\item "s" or "?" = GP-Studio (draft working directory, many users with access)
\item "l" or "??" = GP-Live (private, admin only)
}}

\item{change_this}{A list of values to change in the front matter before rebuilding. Default=NULL. Example: list(Title="Stormy Misty's Foal") would change the title of the lesson to the name of a horsey novel. If gh_proj_name=="all", make sure you set this to something you want to change for everything.}

\item{clean}{Do you want to clean the meta/JSON folder and build everything from scratch? (Gets passed to \code{\link[=compile_unit]{compile_unit()}}). Default=FALSE}

\item{rebuild}{Do you want to force rebuild everything (even if a particular item seems up to date?) default=FALSE (This par gets passed on as rebuild to \code{\link[=compile_unit]{compile_unit()}})}

\item{drive_reconnect}{passed to \code{\link[=update_fm]{update_fm()}} to reconnect google drive IDs}
}
\description{
Function to rebuild a lesson or lessons that found in a virtualized Google Drive for Desktop folder. It will also stage the lesson in the published folder if requested, but will not publish. To do that, you should call \code{\link[=batch_publish]{batch_publish()}}
}
\details{
This assumes that you have Google Drive for Desktop set up and have permissions to access the lesson files.
}
