% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_get_fm.R
\name{batch_get_fm}
\alias{batch_get_fm}
\title{get front-matter for a bunch of projects}
\usage{
batch_get_fm(WD, WD_git = NULL)
}
\arguments{
\item{WD}{working directory; if "?" or "s" supplied, will get key values for all projects in the GP-Studio drive. "??" or "l" will get data for "GP-LIVE"}

\item{WD_git}{default=NULL. If you already know the path to the gp-lessons folder, this is more efficient.}
}
\value{
a list of values for the requested keys for each project on the given drive
}
\description{
Just a map function for \code{\link[=get_fm]{get_fm()}}
}
\concept{batch functions}
