% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogURL.R
\name{catalogURL}
\alias{catalogURL}
\title{catalogURL
Add full url prefix to lesson subdirectory where it will be published at catalog.galacticpolymath.com}
\usage{
catalogURL(relative_ref, repo)
}
\arguments{
\item{relative_ref}{relative path in the current environment}

\item{repo}{the name of the current repository e.g. from \code{\link{whichRepo}}}
}
\description{
catalogURL
Add full url prefix to lesson subdirectory where it will be published at catalog.galacticpolymath.com
}
