% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_lesson_git.R
\name{change_lesson_git}
\alias{change_lesson_git}
\title{change_lesson_git}
\usage{
change_lesson_git(new_name, WD = getwd(), check_current = FALSE)
}
\arguments{
\item{new_name}{The name of the new repo (should be exactly as it is named on \url{https://github.com/galacticpolymath}); It's just the project name--a full URL is not expected.}

\item{WD}{working directory; default= getwd()}

\item{check_current}{do you want to check whether the current listed GitHubPath in the front-matter.yml is good? default=F}
}
\description{
Change the GitHub URL associated with this project (i.e. after you rename the repo on \url{https://github.com/galacticpolymath})
}
\details{
Will run some validation checks and let you know whether the new_name repo exists on the web. May also prompt to run \code{\link[=editor]{editor()}} to initialize some fields.
}
