% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_wd.R
\name{check_wd}
\alias{check_wd}
\title{Check if Working Directory Looks Good}
\usage{
check_wd(simple_out = TRUE, throw_error = TRUE)
}
\arguments{
\item{simple_out}{logical; if TRUE results will be T/F (i.e. TRUE=looks good). If F, returns a list of test_result, parent_folder_name, and project_folder_name. default=T}

\item{throw_error}{logical; Do you want to stop code if error (T) or just throw a warning (F)? default=T}
}
\value{
depends on simple_out
}
\description{
The working directory gets reset sometimes for unknown reasons after running the editor. This is a routine check. Working directory should contain an .Rproj file and be a subfolder of 'Lessons'.
}
