% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_yaml.R
\name{check_yaml}
\alias{check_yaml}
\title{Check front-matter.yml}
\usage{
check_yaml(WD = getwd(), skip = NULL, throw_error = TRUE)
}
\arguments{
\item{WD}{working directory; default= getwd()}

\item{skip}{a character vector of items you want to skip validation for. default=NULL.
Options:
\itemize{
\item locale: Check for existence of locale entry in front-matter.yml
\item gh: Check for existence of GitHubPath entry in front-matter.yml
}}

\item{throw_error}{logical; Do you want to stop code if error (T) or just throw a warning (F)? default=T}
}
\description{
Checks for existence of front-matter.yml in the meta/ subfolder of supplied working directory. Then, validates that certain expected parameters are set
}
