% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearTeachingMatExcelData.R
\name{clearTeachingMatExcelData}
\alias{clearTeachingMatExcelData}
\title{clearTeachingMatExcelData}
\usage{
clearTeachingMatExcelData(tabsToClear = "all")
}
\arguments{
\item{tabsToClear}{a character vector; which tab(s) to delete everything but the first 2 header rows; Options=c("quickPrep_Fb","dL","c_pres","c_handouts","r_pres","r_handouts"). Default is "all".}
}
\description{
Caution! Deletes data for the selected sheet(s) in the teaching-materials.xlsx workbook
}
