% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_fm.R
\name{compile_fm}
\alias{compile_fm}
\alias{fm_compile}
\title{compile_fm}
\usage{
compile_fm(WD = "?", upload = TRUE)

fm_compile(WD = "?", upload = TRUE)
}
\arguments{
\item{WD}{working directory, passed to \code{\link[=parse_wd]{parse_wd()}}}
}
\value{
logical of success
}
\description{
compile front-matter found in /meta/front-matter.yml

alias for for \code{\link[=compile_fm]{compile_fm()}}
}
\details{
Produces the following json outputs in github project \verb{gp-lessons/Lessons/[PROJ_NAME]/JSONs}:
\itemize{
\item header.json
\item overview.json
\item background.json
\item credits.json
\item acknowledgments.json
\item versions.json
}
}
\section{Functions}{
\itemize{
\item \code{fm_compile()}: alias

}}
