% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_json.R
\name{compile_json}
\alias{compile_json}
\title{Compile all the JSONs to create 1 file for the web}
\usage{
compile_json(WD = NULL, WD_git = NULL, destFolder)
}
\arguments{
\item{WD}{working directory, passed to \code{\link[=parse_wd]{parse_wd()}}}

\item{WD_git}{location of gp-lessons github repo. Default=NULL will get this for the current workspace with \code{\link[=get_wd_git]{get_wd_git()}}}

\item{destFolder}{where you want to save the folder; by default in the "WD_git" folder}
}
\value{
tibble of the compiled standards data; a JSON is saved to meta/JSON/UNIT.json
}
\description{
Combine all JSON components into 1 UNIT.json
}
