% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_procedure.R
\name{compile_procedure}
\alias{compile_procedure}
\title{compile_procedure}
\usage{
compile_procedure(WD = getwd(), proc = NULL)
}
\arguments{
\item{WD}{path to the working directory via a virtualized Google Drive for Desktop local path; default = getwd(); if "?" is supplied, this will invoke \code{\link[=pick_lesson]{pick_lesson()}} to get the path for one of the lessons}

\item{proc}{the procedure file if it has already been read in from the teach-it.gsheet, i.e. via \code{\link[=compile_teach_it]{compile_teach_it()}}}
}
\description{
Translate the 'Procedure' tab of the lesson's 'teach-it.gsheet' into JSON format.
}
