% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_standards.R
\name{compile_standards}
\alias{compile_standards}
\title{compile_standards}
\usage{
compile_standards(WD = "?", targetSubj = NULL)
}
\arguments{
\item{WD}{is working directory of the project (useful to supply for shiny app, which has diff. working environment); If you put "?", it will invoke \code{\link[=pick_lesson]{pick_lesson()}}}

\item{targetSubj}{which subject(s) are the focus of the lesson? opts= "math","ela","science","social studies"; default=NULL}
}
\value{
list with 4 objects: $success (did it work?); $input (the input file as a tibble); $compiled (the compiled tibble); $problem_entries (a tibble of entries with 'TBD' or missing values in the "How this aligns..." column). A JSON is saved to the destFolder location.
}
\description{
Does the following:
\enumerate{
\item Compile alignment info from the \code{standards_*.gsheet} in the project's \verb{meta/} folder
\item Output standards.RDS file with the alignment info (this will be read in by \code{\link[=learningEpaulette]{learningEpaulette()}} and \code{\link[=learningChart]{learningChart()}})
}
}
