% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_teach_mat.R
\name{compile_teach_mat}
\alias{compile_teach_mat}
\title{compile_teach_mat}
\usage{
compile_teach_mat(WD = getwd(), teach_it_drib = NULL)
}
\arguments{
\item{WD}{is working directory of the project; easiest way to supply a different lesson is with WD=\code{\link[=pick_lesson]{pick_lesson()}}; default is WD=getwd()}

\item{teach_it_drib}{if you already have the teach-it.gsheet dribble looked up from \code{\link[=drive_find_path]{drive_find_path()}}, passing this object can can save some time; default = NULL}
}
\value{
tibble of the compiled standards data; a JSON is saved to meta/JSON/teaching-materials.json
}
\description{
Compile Teaching Materials from teach-it.gsheet
}
