% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_unit.R
\name{compile_unit}
\alias{compile_unit}
\alias{unit_compile}
\title{compile_unit}
\usage{
compile_unit(
  WD = "?",
  choices,
  current_data,
  clean = FALSE,
  rebuild = NULL,
  drive_reconnect = FALSE
)

unit_compile(
  WD = "?",
  choices,
  current_data,
  clean = FALSE,
  rebuild = NULL,
  drive_reconnect = FALSE
)
}
\arguments{
\item{WD}{is working directory of the project (useful to supply for shiny app, which has diff. working environment); if "?" supplied, will invoke \code{\link[=pick_lesson]{pick_lesson()}}}

\item{choices}{one or more of the following: c("Front Matter","Standards Alignment","Teaching Materials","Procedure","Acknowledgements","Versions"); or "All". If missing, will compile things in the ReadyToCompile entry in front-matter.yml for the WD folder.}

\item{current_data}{the reconciled data including yaml and input from the shiny app environment; if current_data=NULL, read in front-matter.yml}

\item{clean}{delete all JSON files in meta/ and start over? default=FALSE}

\item{rebuild}{if T, rebuild everything; overrides RebuildAllMaterials in front-matter.yml; default= NULL}

\item{drive_reconnect}{passed to \code{\link[=update_fm]{update_fm()}} to reconnect google drive IDs}
}
\value{
success; logical
}
\description{
Compiles selected sections of a lesson (or "all"). Results in a UNIT.json, but files are not staged for publishing. Need to follow with a call to \code{\link[=stage_assets]{stage_assets()}} and \code{\link[=publish]{publish()}} to publish these changes to the web. Tries to use cacheing, but sometimes, may need to run a function twice to get efficient time savings because of delay in modTimes with Google Drive for Desktop.
}
\details{
Combines functionality of:
\itemize{
\item \code{\link[=compile_fm]{compile_fm()}}
\item \code{\link[=compile_standards]{compile_standards()}}
\item \code{\link[=learningEpaulette]{learningEpaulette()}}
\item \code{\link[=compile_json]{compile_json()}}
}

Intended for a single lesson in the current RStudio project. Use \code{\link[=batch_compile]{batch_compile()}} to compile and rebuild more than one lesson (or a single lesson outside the current project).
}
\section{Functions}{
\itemize{
\item \code{unit_compile()}: alias

}}
