% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileAlignment.R
\name{compileAlignment}
\alias{compileAlignment}
\title{compileAlignment}
\usage{
compileAlignment(
  lessonAlignmentMatrix,
  grades = "all",
  fileName = "standards/processedStandards.json"
)
}
\arguments{
\item{lessonAlignmentMatrix}{file location of the lesson alignment matrix XLSX worksheet}

\item{grades}{grade band on alignment matrix worksheet options= "5-6", "7-8", "9-12" or "all"}

\item{fileName}{output file name (including path if you don't want it in the working directory)}
}
\value{
tibble of the compiled standards data; a JSON is saved to standards/processedStandards.json
}
\description{
Compile alignment info from a lesson standards matrix worksheet
}
