% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileStandards.R
\name{compileStandards}
\alias{compileStandards}
\title{compileStandards}
\usage{
compileStandards(
  standardsFile = "meta/standards_GSheetsOnly.xlsx",
  destFolder = "meta/JSON/",
  fileName = "standards.json",
  standardsRef = "standardX"
)
}
\arguments{
\item{standardsFile}{file location of the lesson standards alignment XLSX worksheet; default: "meta/standards_GSheetsOnly.xlsx"}

\item{destFolder}{where you want to save the folder; by default in the "meta/JSON/" folder, 1 level up from the working directory}

\item{fileName}{output file name; default= "standards.json"}

\item{standardsRef}{where do you want to pull down statements and other info for the supplied standards codes? Default="standardX" (i.e. the \href{https://github.com/galacticpolymath/standardX}{standardX repository}); Options= "standardX" or "myFile" (which will use Tab 2 on the supplied XLSX workbook)}
}
\value{
list of the compiled standards data with 3 objects: $input (the input file as a tibble); $compiled (the compiled tibble); $problem_entries (a tibble of entries with 'TBD' or missing values in the "How this aligns..." colum). A JSON is saved to the destFolder location.
}
\description{
Compile alignment info from a lesson standards matrix worksheet
}
