% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileStandards_OLD.R
\name{compileStandards_OLD}
\alias{compileStandards_OLD}
\title{compileStandards_OLD}
\usage{
compileStandards_OLD(
  alignmentMatrixFile = "meta/alignment-matrix.xlsx",
  grades = "5-12",
  destFolder = "meta/JSON/",
  fileName = "standards.json"
)
}
\arguments{
\item{alignmentMatrixFile}{file location of the lesson alignment matrix XLSX worksheet; default: "meta/alignment-matrix.xlsx"}

\item{grades}{grade band on alignment matrix worksheet options= "5-6", "7-8", "9-12" or "5-12"}

\item{destFolder}{where you want to save the folder; by default in the "meta/JSON/" folder, 1 level up from the working directory}

\item{fileName}{output file name; default= "processedStandards.json"}
}
\value{
Compiled standards
}
\description{
Compile alignment info from a lesson standards matrix worksheet. This was the first compilation method, used for the assumptionsMatter lesson, and it's basically garbage. Keeping it here till I can update that lesson for the new workflow.
}
