% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_T_to_check.R
\name{convert_T_to_check}
\alias{convert_T_to_check}
\title{Convert logical values to checks and exes}
\usage{
convert_T_to_check(x, no_match = NA)
}
\arguments{
\item{x}{a logical value or vector of logical values}

\item{no_match}{what to put if there's no match (e.g. non-logical argument)? default=NA}
}
\value{
✓ if TRUE and ✗ if FALSE. If not logical, returns NA or another string supplied by no_match
}
\description{
Takes logical T and F and returns ✓ and ✗, respectively. If not logical, returns NA
}
\details{
It can handle existing ✓ and ✗, so you can run result through convert_T_to_check without problem
}
