% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_dir_selectively.R
\name{copy_dir_selectively}
\alias{copy_dir_selectively}
\title{Copy directory selectively}
\usage{
copy_dir_selectively(source_path, dest_path, ignore_gdrive = TRUE)
}
\arguments{
\item{source_path}{source path on local file system}

\item{dest_path}{full destination path}

\item{ignore_gdrive}{logical; default=TRUE, ignore gdrive files (.gdoc,.gsheet,.gslide,.gform,.gdraw)}
}
\description{
Copy a folder, omitting certain items. Use case: using Gdrive for desktop virtualization, copy a folder, but ignore Gdrive files
}
