% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_updated_files.R
\name{copy_updated_files}
\alias{copy_updated_files}
\title{copy_updated_files}
\usage{
copy_updated_files(paths, dest_folder, clear = FALSE, verbose = TRUE, WD = "?")
}
\arguments{
\item{paths}{vector of file paths you want to check and/or copy}

\item{dest_folder}{directory where you want files to be transferred}

\item{clear}{do you want to delete all files in the destination folder before copying? default=FALSE}

\item{verbose}{print summary table? default=T}

\item{WD}{Working directory. Needed to pass on to inSync.. :/ might want to refactor}
}
\value{
tibble of path1 files and statuses (whether they were updated, added, or are up-to-date)
}
\description{
Copy files to destination only if the file is missing or out of date at the destination. Mainly a helper for \code{\link[=stage_assets]{stage_assets()}}. Not well-designed for general use.
}
