% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyUpdatedFiles.R
\name{copyUpdatedFiles}
\alias{copyUpdatedFiles}
\title{copyUpdatedFiles}
\usage{
copyUpdatedFiles(paths, destFolder, clear = FALSE, verbose = TRUE)
}
\arguments{
\item{paths}{vector of file paths you want to check and/or copy}

\item{destFolder}{directory where you want files to be transferred}

\item{clear}{do you want to delete all files in the destination folder before copying? default=FALSE}

\item{verbose}{print summary table? default=T}
}
\description{
Copy files to destination only if the file is missing or out of date at the destination
}
