% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_and_country_codes.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{country_codes}
\format{
a 249 x 6 tibble
}
\usage{
country_codes
}
\description{
ISO Country codes
}
\details{
These are updated with \code{\link[=update_ISO_codes]{update_ISO_codes()}}
}
\keyword{datasets}
