% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_copy_missing_files.R
\name{drive_copy_missing_files}
\alias{drive_copy_missing_files}
\title{drive_copy_missing_files()}
\usage{
drive_copy_missing_files(
  from_dir,
  to_dir,
  ignore = c(".git", ".github", ".Rproj.user", ".RData", ".Rhistory"),
  prompt_user = TRUE,
  try_harder = TRUE,
  share_anyone = FALSE,
  ...
)
}
\arguments{
\item{from_dir}{Source directory path}

\item{to_dir}{What's the path to the new (clone) directory? Must be a string path for \code{\link[=drive_find_path]{drive_find_path()}} or a dribble, e.g. from \code{\link[googledrive:drive_find]{googledrive::drive_find()}} e.g. "Edu/Lessons/GeneticRescue_sci"}

\item{ignore}{Folders to ignore; default= c(".git", ".github", ".Rproj.user",".RData",".Rhistory")}

\item{prompt_user}{prompt user to continue? default=T; set to F for recursive calls}

\item{try_harder}{do you want copy functions to try again if copying fails? Passed to \code{\link[=catch_err]{catch_err()}}; default=T}

\item{share_anyone}{do you want copied files to be viewable by anyone with the link? Default=FALSE}

\item{...}{Pass other parameters to \code{\link[googledrive:drive_cp]{googledrive::drive_cp()}}}
}
\value{
Dribble of the cloned folder
}
\description{
Copy files missing AND folders from one Google Drive (web) folder to another
}
\details{
Helps with a partial project clone (i.e. if not all files are backed up on GitHub, but they are on GDrive). Makes use of our function \code{\link[=drive_cp_dir]{drive_cp_dir()}} to copy folders.
}
\seealso{
Other Google Drive Functions:
\code{\link[=drive_contents]{drive_contents()}},
\code{\link[=drive_cp_dir]{drive_cp_dir()}},
\code{\link[=drive_find_path]{drive_find_path()}},
\code{\link[=drive_get_info]{drive_get_info()}},
\code{\link[=drive_inSync]{drive_inSync()}},
\code{\link[=drive_new_from_template]{drive_new_from_template()}},
\code{\link[=drive_open]{drive_open()}},
\code{\link[=init_unit_meta]{init_unit_meta()}},
\code{\link[=update_teach_links]{update_teach_links()}}
}
\concept{Google Drive Functions}
