% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_cp_dir.R
\name{drive_cp_dir}
\alias{drive_cp_dir}
\title{Copy a Google Drive Folder}
\usage{
drive_cp_dir(new_dir_name, source_dir, dest_dir, share_anyone = FALSE, ...)
}
\arguments{
\item{new_dir_name}{What's the name of the new (cloned) directory?}

\item{source_dir}{Source directory; must be a string path for \code{\link[=drive_find_path]{drive_find_path()}} or a dribble, e.g. from \code{\link[googledrive:drive_find]{googledrive::drive_find()}}}

\item{dest_dir}{Destination directory; if not supplied, assumes same parent as source_dir (i.e. will be cloned side-by-side)}

\item{share_anyone}{do you want copied files to be viewable by anyone with the link? Default=FALSE}

\item{...}{Pass other parameters to \code{\link[googledrive:drive_cp]{googledrive::drive_cp()}}}
}
\value{
Dribble of the cloned folder
}
\description{
Copies all files (Google Workspace or otherwise) from one destination to another. This is a workaround for the absence of this functionality in the GDrive API. It's accomplished by recursive calls to \code{\link[googledrive:drive_cp]{googledrive::drive_cp()}}.
}
\details{
Throws an error if the new_dir_name already exists in  dest_dir.
}
\seealso{
Other Google Drive Functions:
\code{\link[=drive_contents]{drive_contents()}},
\code{\link[=drive_copy_missing_files]{drive_copy_missing_files()}},
\code{\link[=drive_find_path]{drive_find_path()}},
\code{\link[=drive_get_info]{drive_get_info()}},
\code{\link[=drive_inSync]{drive_inSync()}},
\code{\link[=drive_new_from_template]{drive_new_from_template()}},
\code{\link[=drive_open]{drive_open()}},
\code{\link[=init_unit_meta]{init_unit_meta()}},
\code{\link[=update_teach_links]{update_teach_links()}}
}
\concept{Google Drive Functions}
