% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_find_path.R
\name{drive_find_path}
\alias{drive_find_path}
\title{Find folder or file along Google Drive path}
\usage{
drive_find_path(
  drive_path,
  WD = NULL,
  drive_root = NULL,
  exact_match = TRUE,
  single_result = TRUE,
  checkWD = TRUE,
  raw_results = FALSE,
  ...
)
}
\arguments{
\item{drive_path}{If you provide a dribble, it gets passed right back out. Otherwise, you can give a character string in the form "DRIVE/directory/subdirectory".
\itemize{
\item DRIVE can be "~" or "my drive" to refer to your private google drive (e.g. "~/Folder1/file.png")
\item Path can also be the name of a shared drive (e.g. "GP-Misc")
\item Also supports relative paths (e.g. "../meta"), as long as \strong{WD} or \strong{drive_root}. If \strong{WD} is supplied (this is a full "local" Google Drive for Desktop path piped from \code{\link[=pick_lesson]{pick_lesson()}}). If \strong{drive_root} supplied, it will use that as the parent for getting relative paths.
\item Relative pathing can be significantly faster because each hierarchical call to GDrive API to resolve a folder costs about a second.
\item You can also pass an ID as a text string if you know that already
\item Generally, case SeNsItIvE...but results may vary
}}

\item{WD}{\itemize{
\item a \strong{local} virtualized path to a lesson folder where Google Drive (Web) path will be extracted from front matter. Easiest is to pass "?" whcih will invoke \code{\link[=pick_lesson]{pick_lesson()}}; must use \code{full_path=TRUE} with pick_lesson
\item will be ignored unless relative path provided ("../folder1"), where \strong{WD} will be substituted for ".."
}}

\item{drive_root}{A Google drive path reference (not a local or virtualized path). Will be ignored unless relative path provided ("../folder1"), where drive_root will be substituted for "..". Can be:
\enumerate{
\item a dribble or
\item a Googledrive ID (as a string)
\item drive_root is passed to \code{\link[googledrive:drive_get]{googledrive::drive_get()}}
}}

\item{exact_match}{logical; Do you want an exact match for the file name of the path? (only applies to the final FILENAME part of the path; i.e  'folder/folder/FILENAME_w_different_suffix'); default=TRUE}

\item{single_result}{logical; do you want to force a single result (i.e. throw an error if there is more than one match)?; default=TRUE}

\item{checkWD}{do you want to run \code{\link[=check_wd]{check_wd()}}? default=TRUE; set to FALSE to suppress warnings if for example you're missing teach-it.gsheet or some other item expected to be in a lesson directory}

\item{raw_results}{logical; used for error checking on path resolution. If T, will export the full list of results of found paths, rather than just a tibble of the final target result. Default=FALSE}

\item{...}{other arguments passed to \code{\link[googledrive:drive_find]{googledrive::drive_find()}}}
}
\description{
Simulates Finder/File Explorer functionality by repeated calls to \code{\link[googledrive:drive_find]{googledrive::drive_find()}}. Allows relative paths via hybrid navigation of local virtualized Google Drive for Desktop paths by going to a local lesson directory, pulling up the front-matter.yml and finding the Google ID in order to access the relative path you requested through the Drive API.
}
\examples{
\dontrun{
#ABSOLUTE PATHS
#path to a drive file on your personal Google Drive
drive_find_path("~/folder_in_my_personal_drive/filename")
#path to a network drive folder
(p <- drive_find_path("GP-Studio/Edu/Lessons/geneticrescue_sci"))
#show contents of that drive folder
p \%>\% drive_contents()

#RELATIVE PATH to a particular lesson subfolder
#only works if you have Google Drive for Desktop set up with permissions to GP-Studio
drive_find_path("../assets",pick_lesson()) \%>\% drive_contents
}

}
\seealso{
Other Google Drive Functions:
\code{\link[=drive_contents]{drive_contents()}},
\code{\link[=drive_copy_missing_files]{drive_copy_missing_files()}},
\code{\link[=drive_cp_dir]{drive_cp_dir()}},
\code{\link[=drive_get_info]{drive_get_info()}},
\code{\link[=drive_inSync]{drive_inSync()}},
\code{\link[=drive_new_from_template]{drive_new_from_template()}},
\code{\link[=drive_open]{drive_open()}},
\code{\link[=init_unit_meta]{init_unit_meta()}},
\code{\link[=update_teach_links]{update_teach_links()}}
}
\concept{Google Drive Functions}
