% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_get_comments.R
\name{drive_get_comments}
\alias{drive_get_comments}
\title{Get comments for a Google Drive file}
\usage{
drive_get_comments(drive_file)
}
\arguments{
\item{drive_file}{A google drive dribble. Should be a single file}
}
\value{
A tibble of comments and relevant info
}
\description{
Input a google drive dribble, get a tibble of comments
}
