% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_open_proj_folder.R
\name{drive_open_proj_folder}
\alias{drive_open_proj_folder}
\alias{open_proj_folder_drive}
\title{Open project local (virtualized) project folder}
\usage{
drive_open_proj_folder(WD = "?")

open_proj_folder_drive(WD = "?")
}
\arguments{
\item{WD}{working directory; passed to \code{\link[=parse_wd]{parse_wd()}}.}
}
\description{
Requires Google Drive for Desktop to be set up and to have access to the folder. If it's on GP-LIVE, you may not have access.
}
\section{Functions}{
\itemize{
\item \code{open_proj_folder_drive()}: alias

}}
