% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.R
\name{E}
\alias{E}
\title{E}
\usage{
E(function_name)
}
\description{
Edit function definition in /R subfolder
}
\details{
A convenience wrapper for \code{\link[usethis]{edit_file}} to edit or View a galacticPubs function. If
}
