% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_script.R
\name{edit_script}
\alias{edit_script}
\title{Create/Edit a script for a GP unit}
\usage{
edit_script(script_name = NULL, WD = "?")
}
\arguments{
\item{script_name}{text string to name the script. No need to add .R suffix.}

\item{WD}{working directory. Default '?'}
}
\value{
logical; if succeeds, should open a new file and silently return TRUE
}
\description{
If the script name is not found, it will create one in the gp-lessons/lessons/(working directory) aka wd_git for the unit
}
