% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efx.R
\name{efx}
\alias{efx}
\title{efx}
\usage{
efx(function_name)
}
\arguments{
\item{function_name}{name of a galacticPubs function}
}
\description{
Edit function definition in /R subfolder
}
\details{
A convenience wrapper for \code{\link[usethis]{edit_file}} to edit or View a galacticPubs function. If in the galacticPubs dev project, it will edit; otherwise view.
}
