% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_md_links.R
\name{expand_md_links}
\alias{expand_md_links}
\title{expand_md_links}
\usage{
expand_md_links(md, WD)
}
\arguments{
\item{md}{a markdown string or vector}

\item{WD}{the Working Directory}
}
\description{
Make markdown links into full paths to GP catalog. Supports 2 types of links:
\itemize{
\item Named Link: \verb{![Link Text](filename.ext)}
\item Unnamed Link: \code{{filename.ext}}
}
}
\details{
Ex:
\itemize{
\item "See \verb{![image1](img.png)} for details" becomes ->
\item "See \verb{![image1](https://catalog.galacticpolymath.com/ThisProject/img.png)} for details"
Ex: 2:
\item "\code{{img.png}}" becomes ->
\item \verb{\{https://catalog.galacticpolymath.com/ThisProject/img.png\}}
}
}
\seealso{
Other markdown:
\code{\link[=fixAnchorLinks]{fixAnchorLinks()}},
\code{\link[=parseGPmarkdown]{parseGPmarkdown()}}
}
\concept{markdown}
