% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandMDLinks.R
\name{expandMDLinks}
\alias{expandMDLinks}
\title{expandMDLinks}
\usage{
expandMDLinks(md, repo)
}
\arguments{
\item{md}{a markdown string or vector}

\item{repo}{the name of the current repository e.g. from \code{\link{whichRepo}}}
}
\description{
Make markdown links into full paths to GP catalog
}
\details{
Ex: "See ![image1](img.png) for details" becomes ->
"See ![image1](https://catalog.galacticpolymath.com/ThisRepository/img.png) for details"
}
\seealso{
[parseGPmarkdown()]

Other markdown: 
\code{\link{fixAnchorLinks}()},
\code{\link{parseGPmarkdown}()}
}
\concept{markdown}
