% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixAnchorLinks.R
\name{fixAnchorLinks}
\alias{fixAnchorLinks}
\title{fixAnchorLinks}
\usage{
fixAnchorLinks(md)
}
\arguments{
\item{md}{a markdown string or vector}
}
\description{
Converts \verb{(blah)[#tag]} notation to react component: \verb{<AnchorLink href="#tag">blah</AnchorLink>}
}
\seealso{
\code{\link[=parseGPmarkdown]{parseGPmarkdown()}}

Other markdown:
\code{\link[=expand_md_links]{expand_md_links()}},
\code{\link[=parseGPmarkdown]{parseGPmarkdown()}}
}
\concept{markdown}
