% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_restore.R
\name{fm_restore}
\alias{fm_restore}
\title{Restore front-matter from a backup JSON file in WD Git}
\usage{
fm_restore(WD = "?")
}
\arguments{
\item{WD}{is working directory of the project; easiest way to supply a different lesson is with "?", which will invoke \code{\link[=parse_wd]{parse_wd()}}; default is WD=getwd()}
}
\value{
success; logical
}
\description{
This function restores the front-matter of a specified file from a backup JSON file created with \code{\link[=fm_backup]{fm_backup()}} stored in the WD Git repository.
}
