% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fm_names.R
\name{get_fm_names}
\alias{get_fm_names}
\title{get_fm_names}
\usage{
get_fm_names(sort_names = FALSE)
}
\arguments{
\item{sort_names}{boolean; sort the names in alphabetical order; default = FALSE sorts them in the order of the template, which is kind of based on importance}
}
\value{
names of the template front matter
}
\description{
Get the names (keys) for the front-matter from the template built into galacticPubs. (May differ slightly from a specific front-matter.yml if you've added something manually or not updated recently).
}
