% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gp_api_token.R
\name{get_gp_api_token}
\alias{get_gp_api_token}
\alias{gp_api_get_token}
\title{Get a JWT token for GP API}
\usage{
get_gp_api_token()

gp_api_get_token()
}
\description{
Will get a token from the R environment if available and test it with an empty query to GP API. If missing or expired, will attempt to renew through Google Oauth flow.
}
\section{Functions}{
\itemize{
\item \code{gp_api_get_token()}: alias

}}
