% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_page_metrics.R
\name{get_page_metrics}
\alias{get_page_metrics}
\title{Retrieve Google Analytics Metrics for Specific Page Paths}
\usage{
get_page_metrics(
  property_id,
  paths = "/",
  limit = -1,
  metrics = c("screenPageViews", "averageSessionDuration", "sessionsPerUser",
    "bounceRate", "newUsers", "totalUsers"),
  dimensions = c("pagePath", "sessionSource", "sessionSourceMedium", "country"),
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{property_id}{A character string specifying the Google Analytics View ID.}

\item{paths}{A character vector of page paths to filter the results.}

\item{limit}{passed to \code{\link[googleAnalyticsR:ga_data]{googleAnalyticsR::ga_data()}}; default= -1 (all rows, overriding function default of 100)}

\item{metrics}{metrics to provide; default=c("screenPageViews", "averageSessionDuration", "sessionsPerUser", "bounceRate","newUsers", "totalUsers")}

\item{dimensions}{what variables to categorize data by? default= c("pagePath", "sessionSource", "sessionSourceMedium", "country")}

\item{start_date}{A character string specifying the start date (default: "30daysAgo").}

\item{end_date}{A character string specifying the end date (default: "yesterday").}
}
\value{
A data frame containing Google Analytics metrics for the specified paths.
}
\description{
This function fetches key Google Analytics metrics for specified page paths, including
the number of visits, session duration, traffic sources, user locations, and other relevant
statistics. It helps analyze the engagement and reach of specific pages.
}
\examples{
# Example usage (replace "YOUR_property_id" with an actual Google Analytics View ID)
# paths <- c("/example-page", "/another-page")
# data <- get_page_metrics(property_id = "YOUR_property_id", paths = paths)
# print(data)
}
\seealso{
\href{https://developers.google.com/analytics/devguides/reporting/data/v1/api-schema}{Google Analytics Data API Schema}
}
