% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unit_path.R
\name{get_unit_path}
\alias{get_unit_path}
\title{Get local (virtualized Google Drive for Desktop) path to unit}
\usage{
get_unit_path(str, pull_path = TRUE)
}
\arguments{
\item{str}{a string that matches the name of a project in one of the shared drives GP-LIVE or GP-Studio}
}
\value{
path(s) for the match(es) to the string you supplied
}
\description{
Similar to \code{\link[=pick_lesson]{pick_lesson()}}, except it expects you to supply a ShortTitle or exact project title with _locale suffix
}
