% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGalactic.R
\name{ggGalactic}
\alias{ggGalactic}
\title{ggGalactic}
\usage{
ggGalactic(
  font = "Montserrat",
  regular.wt = 400,
  bold.wt = 700,
  font.cex = 1,
  plot.margin = ggplot2::margin(t = 10, r = 10, b = 10, l = 10)
)
}
\arguments{
\item{font}{Google font to use, "Montserrat" by default}

\item{regular.wt}{font weight for regular font style}

\item{bold.wt}{font weight for bold text}

\item{font.cex}{a simple multiplier for scaling all text}

\item{plot.margin}{easy access to ggplot margins}
}
\description{
Galactic Polymath stylings for ggplot2
}
