% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_add_this_repo.R
\name{gh_add_this_repo}
\alias{gh_add_this_repo}
\title{Add this repository to galacticpolymath github}
\usage{
gh_add_this_repo()
}
\description{
This functino expects that:
\itemize{
\item You have \href{https://cli.github.com/manual/installation}{gh cli} (github command line) installed
\item You're set up with git in Rstudio
\item You have credentials to write to galacticpolymath github repos
}
}
