% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_delete_repo.R
\name{gh_delete_repo}
\alias{gh_delete_repo}
\title{Delete a Galactic Polymath Lesson Repo}
\usage{
gh_delete_repo()
}
\description{
This expects you have git \emph{and} gh set up to work in terminal and that you have credentials to modify and delete a repository at https://github.com/galacticpolymath
}
