% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_remove_from_GPcatalog.R
\name{gh_remove_from_GPcatalog}
\alias{gh_remove_from_GPcatalog}
\title{Check for and remove project from GP Catalog}
\usage{
gh_remove_from_GPcatalog(gh_proj_name)
}
\arguments{
\item{gh_proj_name}{The (OLD) unique project title of this lesson as it is named on \url{https://github.com/galacticpolymath}.}
}
\value{
logical; TRUE= success, FALSE= fail, NA= project not found
}
\description{
Useful for when you rename a github repo of a lesson project and you have an orphaned catalog.galacticpolymath.com folder.
}
