% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_api_names.R
\name{gp_api_names}
\alias{gp_api_names}
\title{Fields for GP Mini-Unit records}
\usage{
gp_api_names()
}
\description{
Returns a vector of top-level sections of LESSON.json records in GP's MongoDB catalog.
}
