% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_api_query.R
\name{gp_api_query}
\alias{gp_api_query}
\alias{query_gp_api}
\title{Get info for Galactic Polymath units}
\usage{
gp_api_query(
  keys = NULL,
  numID = NULL,
  output_tibble = TRUE,
  dev = FALSE,
  id = NULL,
  sort_by = "numID",
  verbosity = NULL
)

query_gp_api(
  keys = NULL,
  numID = NULL,
  output_tibble = TRUE,
  dev = FALSE,
  id = NULL,
  sort_by = "numID",
  verbosity = NULL
)
}
\arguments{
\item{keys}{character vector; which front-matter keys do you want from lessons? default:NULL; use "basic" as shorthand for c("numID","_id","Title"). See all options with \code{\link[=get_fm_names]{get_fm_names()}}}

\item{numID}{is a vector of numIDs for unit(s) you want. default=NULL returns all units}

\item{output_tibble}{return values as a "tibble"? otherwise, list; default=TRUE}

\item{dev}{logical; if FALSE (default), gets catalog from the production gp-catalog. Otherwise, from the dev catalog.}

\item{id}{is a vector of \verb{_id}s for unit(s) you want. default=NULL returns all units}

\item{sort_by}{a character giving the column name to sort by. Default="numID"}

\item{verbosity}{verbosity passed to \code{\link[httr2:req_perform]{httr2::req_perform()}}}
}
\value{
list of results or tbl_json
}
\description{
Constructs a query and requests selected information from the from the galacticpolymath.com/api
}
\section{Functions}{
\itemize{
\item \code{query_gp_api()}: Alias for gp_api_query

}}
\seealso{
Other GP API:
\code{\link[=gp_api_get_token]{gp_api_get_token()}},
\code{\link[=gp_api_query_users]{gp_api_query_users()}},
\code{\link[=gp_api_unit_delete]{gp_api_unit_delete()}},
\code{\link[=gp_api_unit_insert]{gp_api_unit_insert()}},
\code{\link[=gp_api_unit_replace]{gp_api_unit_replace()}}
}
\concept{GP API}
