% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_api_unit_insert.R
\name{gp_api_unit_insert}
\alias{gp_api_unit_insert}
\title{Insert a new GP unit into the database}
\usage{
gp_api_unit_insert(WD = "?", dev = TRUE, verbosity = 1)
}
\arguments{
\item{WD}{working directory, passed to \code{\link[=parse_wd]{parse_wd()}};default="?"}

\item{dev}{logical; if FALSE, gets catalog from the production gp-catalog. Otherwise, from the dev catalog. TRUE (default) will apply to both dev and prod catalogs.}

\item{verbosity}{passed to \code{\link[httr2:req_perform]{httr2::req_perform()}}; default=1}
}
\value{
success (logical)
}
\description{
Insert (i.e. add) a new unit to MongoDB using the GP API.
}
\seealso{
Other GP API:
\code{\link[=gp_api_get_token]{gp_api_get_token()}},
\code{\link[=gp_api_query]{gp_api_query()}},
\code{\link[=gp_api_query_users]{gp_api_query_users()}},
\code{\link[=gp_api_unit_delete]{gp_api_unit_delete()}},
\code{\link[=gp_api_unit_replace]{gp_api_unit_replace()}}
}
\concept{GP API}
