% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_api_unit_replace.R
\name{gp_api_unit_replace}
\alias{gp_api_unit_replace}
\title{Replace a GP unit in the database}
\usage{
gp_api_unit_replace(
  WD = "?",
  prompt_user = TRUE,
  dev = NULL,
  verbosity = 1,
  print_output = TRUE
)
}
\arguments{
\item{WD}{working directory, passed to \code{\link[=parse_wd]{parse_wd()}};default="?"}

\item{prompt_user}{logical; ask user before deleting and replacing the unit? default=TRUE}

\item{dev}{logical; default (NULL) modifies both production and dev gp-catalogs. FALSE modifies the production gp-catalog. TRUE modifies only the dev catalog. supplying c(TRUE,TRUE) is the same as NULL}

\item{verbosity}{passed to \code{\link[httr2:req_perform]{httr2::req_perform()}}; default=1}

\item{print_output}{logical; print result to user? default=TRUE}
}
\value{
success (logical)
}
\description{
Completely deletes and re-inserts a UNIT.json for the mini-unit on the MongoDB database (aka gp-catalog). There are 2 catalogs:
\itemize{
\item dev (not public-facing; viewable through the dev.galacticpolymath.com)
\item production (public facing; live on galacticpolymath.com)
}
}
\details{
Which catalog you're modifying is controlled by the dev parameter. Default is to modify both (so they don't get out of sync). Shallow wrapper for internal functions \code{\link[=gp_api_unit_delete]{gp_api_unit_delete()}} and \code{\link[=gp_api_unit_insert]{gp_api_unit_insert()}}
}
\seealso{
Other GP API:
\code{\link[=gp_api_get_token]{gp_api_get_token()}},
\code{\link[=gp_api_query]{gp_api_query()}},
\code{\link[=gp_api_query_users]{gp_api_query_users()}},
\code{\link[=gp_api_unit_delete]{gp_api_unit_delete()}},
\code{\link[=gp_api_unit_insert]{gp_api_unit_insert()}}
}
\concept{GP API}
