% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_footer.R
\name{gp_footer}
\alias{gp_footer}
\title{gp_footer}
\usage{
gp_footer(
  obj,
  caption = NULL,
  x = 0.015,
  y = 0.05,
  text_size = 9,
  add_border = TRUE,
  border_col_graph = "#363636",
  border_col_foot = "#363636",
  border_width_graph = 3,
  border_width_foot = 1,
  text_color = "#363636",
  data_attrib = NULL,
  data_attrib_prefix = "Data: ",
  data_attrib_scale = 0.9,
  data_attrib_x = 0.75,
  logo = "black",
  logo_scale = 0.7,
  fill_col = "gray90",
  show_plot = TRUE,
  clear_cache = FALSE
)
}
\arguments{
\item{obj}{ggplot object}

\item{caption}{plot caption}

\item{x}{x how much to inset caption as a fraction; default=0.01}

\item{y}{y height of box as a fraction; default=0.08;}

\item{text_size}{text size; default=9;}

\item{add_border}{logical; add border around main plot? default=TRUE}

\item{border_col_graph}{color of border around main plot. default="#363636"; put "transparent" to remove border}

\item{border_col_foot}{color of border around plot footer. default="#363636"; put "transparent" to remove border}

\item{border_width_graph}{width of border around main plot. default=1}

\item{border_width_foot}{width of border around footer. default=1}

\item{text_color}{text color; default="#363636" aka Galactic Black}

\item{data_attrib}{concise citation for data source (will be smaller font and follow "Data: ")}

\item{data_attrib_prefix}{How to prefix data_attrib text. By default "Data: "}

\item{data_attrib_scale}{default=0.6 from 0 to 1, how much smaller to scale attribution text?}

\item{data_attrib_x}{where do you want to place the right edge of the data attribution text (0 to 1); default= 0.75}

\item{logo}{default="black"; which GP logo do you want to use?}

\item{logo_scale}{on a scale of 0 to 1 (npc units), how high to make the logo in footer? default=0.7}

\item{fill_col}{fill color for caption at bottom; default is "gray90"}

\item{show_plot}{logical; plot resulting figure? default=T}

\item{clear_cache}{logical; clear cached logo file? default=FALSE}
}
\description{
Add a Galactic Polymath footer to the bottom of ggplot objects
}
