% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpPal.R
\docType{data}
\name{gpPal}
\alias{gpPal}
\title{Galactic Polymath color stylings}
\format{
A list with 4 items. Each item is a data frame with color names, hex and rgb codes
}
\usage{
gpPal
}
\description{
Easy access to GP colors and shapes
}
\keyword{datasets}
