% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpsave.R
\name{gpsave}
\alias{gpsave}
\title{gpsave}
\usage{
gpsave(filename, width = 6, height = 3.38, dpi = 300, ...)
}
\arguments{
\item{filename}{expects filename, with file extension (e.g. "plot.png"); can also include subfolder (e.g. "newfolder/plot.png")}

\item{width}{plot width in inches (default= 6)}

\item{height}{plot height in inches (default= 3.38, i.e. 16:9 aspect ratio)}

\item{dpi}{dots per inch resolution (default= 300); note changing this will unfortunately also change the output relative text size}

\item{...}{other parameters from [ggplot2::ggsave()]}
}
\description{
Wrapper for [ggplot2::ggsave()] with galacticPubs defaults
}
\details{
Most important thing it does is default to saving in the assets/R_outputs/ folder and save with reasonable dimensions that maintain large, readable text size of graph labels
}
