% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hard_left_join.R
\name{hard_left_join}
\alias{hard_left_join}
\title{hard_left_join}
\usage{
hard_left_join(df1, df2, by, as_character = FALSE)
}
\arguments{
\item{df1}{the data frame (or tibble) whose structure you want to preserve}

\item{df2}{the data frame with new data you want to crush over df1 (destroying matching cells in df1)}

\item{by}{a single column name you want to use to match df1 and df2 values}

\item{as_character}{logical; do you want to override classes and make everything characters. This will avoid errors of matching incompatible types, but may have unexpected results for some data sets. Default=F}
}
\value{
A data frame that has the structure of df1, with matching ids and columns replaced with df2 values
}
\description{
Merge one dataframe into another, replacing data frame 1's values with matching values from data frame 2 without a bunch of .x .y column duplication BS. This is a destructive function, so back up your data and make sure you're getting the expected output.
}
