% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_fork_lesson.R
\name{how_fork_lesson}
\alias{how_fork_lesson}
\title{How to "fork" a lesson to make a new language version?}
\usage{
how_fork_lesson()
}
\description{
Instructions for the steps needed to copy a whole galacticPubs lesson
}
