% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_fm.R
\name{init_fm}
\alias{init_fm}
\title{Initialize Front Matter}
\usage{
init_fm(WD = getwd())
}
\arguments{
\item{WD}{the working directory of the lesson project; default=getwd()}
}
\value{
path to front-matter & a message if a new front-matter.yml file is created
}
\description{
Will run \code{\link[=check_fm]{check_fm()}} and if meta/front-matter.yml not found, it will create this file from the galacticPubs template.
}
