% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_galacticPubs.R
\name{init_galacticPubs}
\alias{init_galacticPubs}
\title{init_galacticPubs()}
\usage{
init_galacticPubs()
}
\value{
table of galacticPubs-related environmental variables, their values, and whether they were successfully set
}
\description{
Set up environmental variables to store local user credentials and paths to resources. Specifically:
\itemize{
\item the local path to GP-Studio/Edu/Lessons, virtualized by \href{https://www.google.com/drive/download}{Google Drive for Desktop}
\item the gp-lessons repository
\item other things
}
}
