% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_gcs.R
\name{init_gcs}
\alias{init_gcs}
\alias{gcs_init}
\title{init_gcs}
\usage{
init_gcs()

gcs_init()
}
\value{
logical; did we succeed in connecting to the cloud?
}
\description{
Initialize connection to Google Cloud Storage galacticPubs buckets
}
\details{
This function uses {googleCloudStorageR} functions to read/write image files to our asset server. To authenticate, it needs access to a JSON file stored in the GP-Dev folder. To have access to this, you need to have Google Drive for Desktop installed and to be in the galacticPubs-user Google Group with our organization. You also need a galacticpolymath.com email address.
}
\section{Functions}{
\itemize{
\item \code{gcs_init()}: alias for init_gcs

}}
