% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_lesson_meta.R
\name{init_lesson_meta}
\alias{init_lesson_meta}
\alias{lesson_init_meta}
\title{init_lesson_meta()}
\usage{
init_lesson_meta(WD = getwd())

lesson_init_meta(WD = getwd())
}
\arguments{
\item{WD}{the working directory for the virtualized lesson path; default=getwd() if you're working in the lesson's .Rproj. Otherwise use \code{\link[=pick_lesson]{pick_lesson()}}}
}
\description{
Add galacticPubs Google Workspace working documents to a given lesson's \emph{/meta} folder. You \emph{must have access to GP-Workshop shared drive}.
}
\details{
WD is used for hybrid file path navigation using virtualized paths on Google Drive for desktop to reach the Edu/Lessons/meta/front-matter.yml for the project of interest. Then, using \code{\link[=get_fm]{get_fm()}} to lookup the GoogleDrive ID for the cloud version of this folder, files will then be copied from the template folder to this project folder using the google web API. All of this is necessary because Google Drive for Desktop doesn't allow copying of Google Documents.
}
\section{Functions}{
\itemize{
\item \code{lesson_init_meta()}: 

}}
\seealso{
Other Google Drive Functions: 
\code{\link{drive_contents}()},
\code{\link{drive_copy_missing_files}()},
\code{\link{drive_cp_dir}()},
\code{\link{drive_find_path}()},
\code{\link{drive_new_from_template}()},
\code{\link{drive_open}()},
\code{\link{drive_parse_name}()}
}
\concept{Google Drive Functions}
