% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_lesson_meta.R
\name{init_unit_meta}
\alias{init_unit_meta}
\alias{unit_init_meta}
\title{init_unit_meta()}
\usage{
init_unit_meta(WD = "?", overwrite = FALSE, template = NULL, override = FALSE)

unit_init_meta(WD = "?", overwrite = FALSE, template = NULL, override = FALSE)
}
\arguments{
\item{WD}{the working directory for the virtualized lesson path; default=getwd() if you're working in the lesson's .Rproj. If "?" is supplied, it will invoke \code{\link[=pick_lesson]{pick_lesson()}}}

\item{overwrite}{logical; Do you want to overwrite target if exact file name found? Does not get passed to \code{\link[googledrive:drive_cp]{googledrive::drive_cp()}} because the way this works is stupid and slow. Instead, we check using virtualized Google Drive for Desktop paths and will overwrite the \emph{exact} file name if T. Default= FALSE.}

\item{template}{which template do you want to copy; default=NULL copies all; options= "standards" and "teach-it"}

\item{override}{boolean; do you want to force copying this, overriding the overwrite logic (which only kinda works); default= FALSE}
}
\value{
logical of success; T=template gsheets copied to meta/ and front-matter updated with \code{\link[=update_fm]{update_fm()}}
}
\description{
Add galacticPubs Google Workspace working documents to a given lesson's \emph{/meta} folder. You \emph{must have access to GP-Studio shared drive}.
}
\details{
WD is used for hybrid file path navigation using virtualized paths on Google Drive for desktop to reach the Edu/Lessons/meta/front-matter.yml for the project of interest. Then, using \code{\link[=get_fm]{get_fm()}} to lookup the GoogleDrive ID for the cloud version of this folder, files will then be copied from the template folder to this project folder using the google web API. All of this is necessary because Google Drive for Desktop doesn't allow copying of Google Documents. It will check existence of templates in target directory using virtualized system for efficiency and to prevent duplications.
}
\section{Functions}{
\itemize{
\item \code{unit_init_meta()}: alias

}}
\seealso{
Other Google Drive Functions:
\code{\link[=drive_contents]{drive_contents()}},
\code{\link[=drive_copy_missing_files]{drive_copy_missing_files()}},
\code{\link[=drive_cp_dir]{drive_cp_dir()}},
\code{\link[=drive_find_path]{drive_find_path()}},
\code{\link[=drive_get_info]{drive_get_info()}},
\code{\link[=drive_inSync]{drive_inSync()}},
\code{\link[=drive_new_from_template]{drive_new_from_template()}},
\code{\link[=drive_open]{drive_open()}},
\code{\link[=update_teach_links]{update_teach_links()}}
}
\concept{Google Drive Functions}
