% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\name{is_empty}
\alias{is_empty}
\title{is_empty}
\usage{
is_empty(x)
}
\arguments{
\item{x}{any value or vector}
}
\value{
TRUE (if empty) or FALSE (if not empty)
}
\description{
Logical test for all the different versions of empty values, i.e.:
\itemize{
\item NULL
\item ""
\item NA
\item "\verb{\\\\n}"
\item list()
\item length(x)==0
}
}
\details{
Works on values, vectors, or a flat list
}
\examples{
is_empty("")
is_empty(3)
is_empty(c("",NA,NULL))
is_empty(list(a="",b=NA,c=NULL))
is_empty(list(a="",b=NA,c="nonempty cell"))
}
