% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\docType{data}
\name{jobs}
\alias{jobs}
\title{jobs}
\format{
a list with info about where the data comes from and what years it covers and the data itself. Contains basic info about job projections from the US Bureau of Labor Statistics.
}
\usage{
jobs
}
\description{
JobViz data for aligning lessons to jobs
}
\keyword{datasets}
