% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_rmd.R
\name{join_rmd}
\alias{join_rmd}
\title{join_rmd}
\usage{
join_rmd(dir, WD = "?", filename = "lesson_plan.Rmd")
}
\arguments{
\item{dir}{relative path to the directory}

\item{WD}{working directory; default= getwd()}

\item{filename}{name of resulting merged RMD; default<-"lesson_plan.Rmd"}
}
\description{
Helper for \code{\link[=make_printable]{make_printable()}} Combines rmd files in the given directory; specifically, take the header.yml, concatenate with other Rmd files in the directory, in alphabetical order
}
