% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_and_country_codes.R
\docType{data}
\name{language_codes}
\alias{language_codes}
\title{language_codes}
\format{
a 486 x 4 tibble
}
\usage{
language_codes
}
\description{
ISO Language codes
}
\details{
These are updated with \code{\link[=update_ISO_codes]{update_ISO_codes()}}
}
\keyword{datasets}
