% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningChart.R
\name{learningChart}
\alias{learningChart}
\title{learningChart}
\usage{
learningChart(
  compiledAlignment,
  targetSubj = NULL,
  caption,
  centralText = "grades\\n5-12",
  centralTextSize = 3.7,
  fileName = "assets/GP_Learning_Chart.png",
  ...
)
}
\arguments{
\item{compiledAlignment}{the output of \code{\link{compileAlignment}}}

\item{targetSubj}{which subject(s) is (are) the focus of the lesson? opts= "math","ela","science","social studies"}

\item{caption}{quoted text you want to go at the bottom of the chart}

\item{centralText}{what you want at the center of the plot ("grades 5-12" by default)}

\item{centralTextSize}{multiplier for font size of centralText}

\item{fileName}{expects somefilename.png}

\item{...}{other arguments passed to \code{\link[grDevices]{png}}}
}
\value{
the learning chart plot object (grid format); the file is saved to assets/GP_Learning_Chart.png by default
}
\description{
Make a GP Learning Chart
}
