% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningEpaulette.R
\name{learningEpaulette}
\alias{learningEpaulette}
\title{learningEpaulette}
\usage{
learningEpaulette(
  compiledAlignment,
  targetSubj = NULL,
  vertSpacing = c(1, 1, 1, 1),
  fileName = "STEAM_epaulette.png"
)
}
\arguments{
\item{compiledAlignment}{the output of \code{\link{compileAlignment}}}

\item{targetSubj}{which subject(s) is (are) the focus of the lesson? opts= "math","ela","science","social studies"}

\item{vertSpacing}{4 value vector ranging from 0 to 1 for manipulating label spacing}

\item{fileName}{expects somefilename.png for ggsave output image file}
}
\description{
Create a Galactic Polymath Learning Epaulette which is a special kind of mosaic plot showing proportion of lesson by subject
}
