% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesson_delete_everywhere.R
\name{lesson_delete_everywhere}
\alias{lesson_delete_everywhere}
\title{Delete a GP lesson \emph{everywhere}}
\usage{
lesson_delete_everywhere(gh_proj_name)
}
\arguments{
\item{gh_proj_name}{name of the folder and associated Galactic Polymath Github project repo}
}
\value{
tibble showing whether project was found and/or deleted. NAs propagated if canceled by user
}
\description{
Deletes lesson from the following places:
\itemize{
\item Google Drive 'Edu/Lessons/...'
\item GitHub (if there is an associated record)
\item GP Catalog (if the lesson has already been published)
}
}
\details{
User will be prompted to continue.
}
