% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesson_go_live.R
\name{lesson_go_live}
\alias{lesson_go_live}
\title{lesson_go_live}
\usage{
lesson_go_live(WD = "?")
}
\arguments{
\item{WD}{is working directory of the project; easiest way to supply a different lesson is with "?", which will invoke \code{\link[=pick_lesson]{pick_lesson()}}; default is WD=getwd()}
}
\description{
A helper function for \code{\link[=organize_teach_it]{organize_teach_it()}}, which is part of \code{\link[=compile_teach_it]{compile_teach_it()}}. Moves/organizes lessons listed as "live" for final public access.
Does the following:
\enumerate{
\item Move lesson project directory from GP-Studio to GP-LIVE (Making it admin-access only)
\item Move teaching-materials/ content from GP-LIVE to GalacticPolymath Shared Drive (Read-Only access)
\item Create shortcut to teaching-materials/ in GP-LIVE
\item Give project a full "MediumTitle" i.e. more descriptive than typical working title of project
\item Update teach-it_*.gsheet DownloadLinks to reflect these file rearrangements
\item Make the following updates to front-matter:
\itemize{
\item PublicationStatus: 'live'
\item GdriveHome: 'GP-LIVE'
}
}
}
