% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesson_init_helper.R
\name{lesson_init_helper}
\alias{lesson_init_helper}
\title{Run lesson_init_helper Shiny app}
\usage{
lesson_init_helper()
}
\description{
Lets you set the following values for your lesson:
\itemize{
\item ShortName
\item Language
\item Country
\item N_lessons
}
}
\details{
Selections are passed to the global environment and interpreted by \code{\link[=lesson_init]{lesson_init()}}
}
