% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lessonBuildSummary.R
\name{lessonBuildSummary}
\alias{lessonBuildSummary}
\title{lessonBuildSummary}
\usage{
lessonBuildSummary()
}
\description{
Create a "Lesson Build Summary.html"
}
\details{
This compiles a lot of information about the lesson, including directory structure.
}
