% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lessons_get_path.R
\name{lessons_get_path}
\alias{lessons_get_path}
\title{Get path to Edu/Lessons folder containing GP projects}
\usage{
lessons_get_path()
}
\description{
Retrieves an environmental variable set with \code{\link[=set_drive_local_credentials]{set_drive_local_credentials()}}
}
