% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locale_picker.R
\name{locale_picker}
\alias{locale_picker}
\title{Run locale_picker Shiny app}
\usage{
locale_picker()
}
\description{
Lets you pick Language and (optional) Country to create a locale. Selections are passed to the global environment
}
