% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumpItems.R
\name{lumpItems}
\alias{lumpItems}
\title{lumpItems}
\usage{
lumpItems(items, item.labs, list.obj, new.name)
}
\arguments{
\item{items}{a vector of names from the list that you want to concatenate}

\item{item.labs}{the labels you want to have boldly concatenated in the text. For example, give "Name(s):" and it will be bolded *Name(s): * and prepended to Item 1 text; supplying "" will not prepend any label to that item}

\item{list.obj}{is the list object you want to work with}

\item{new.name}{is the replacement name for combined text}
}
\value{
The condensed list
}
\description{
Helper function to concatenate multiple list items into a single list item
}
\details{
Specifically designed for lumping separate markdown/YAML entries (which are separated for end user continuity)
into a single list item for the JSON output for the web; output is the supplied list.obj, with items removed and lumped with new.name
}
