% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_printable.R
\name{make_lesson_pdf}
\alias{make_lesson_pdf}
\title{make_lesson_pdf}
\usage{
make_lesson_pdf(WD = getwd(), dest_folder, rebuild = NULL)
}
\arguments{
\item{WD}{is working directory of the project; default= getwd()}

\item{dest_folder}{where do you want template and knitted files to go? default= "assets/lesson-plan-markdown"}

\item{rebuild}{do you want to delete files in the dest_folder and start over? default=NULL}
}
\description{
Knit a concise, formatted PDF version of the lesson plan
}
