% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shareable_assets.R
\name{make_shareable_assets}
\alias{make_shareable_assets}
\title{make_shareable_assets}
\usage{
make_shareable_assets(WD = "?", open_file = TRUE)
}
\arguments{
\item{WD}{"?"}

\item{open_file}{logical; open resulting HTML file? Default=TRUE}
}
\value{
success
}
\description{
Make an HTML file with shareable assets for creating a unit Press Kit for clients
}
