% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesson_go_live.R, R/make_teaching_mat_public.R
\name{make_teaching_mat_public}
\alias{make_teaching_mat_public}
\title{make_teaching_mat_public}
\usage{
make_teaching_mat_public(WD = "?")

make_teaching_mat_public(WD = "?")
}
\arguments{
\item{WD}{is working directory of the project; easiest way to supply a different lesson is with "?", which will invoke \code{\link[=pick_lesson]{pick_lesson()}}; default is WD=getwd()}
}
\description{
A helper function for \code{\link[=organize_teach_it]{organize_teach_it()}}, which is part of \code{\link[=compile_teach_it]{compile_teach_it()}}. Moves/organizes teaching-materials folder to the GalacticPolymath/ shared drive, where it is viewable and copyable by GP+ members.
Does the following:
\enumerate{
\item Move teaching-materials/ content from the WD (project working directory on GP-Studio) to GalacticPolymath Shared Drive (Read-Only access)
\item Create shortcut to teaching-materials/ in place of the moved folder
\item Rename teaching-materials with a full "MediumTitle" i.e. more descriptive than typical working title of project
\item Update teach-it_*.gsheet DownloadLinks to reflect these file rearrangements
\item Make the following updates to front-matter:
\itemize{
\item GdriveTeachMatPath: 'GalacticPolymath/MediumTitle'
}
}

A helper function for \code{\link[=organize_teach_it]{organize_teach_it()}}, which is part of \code{\link[=compile_teach_it]{compile_teach_it()}}. Moves/organizes teaching-materials folder to the GalacticPolymath/ shared drive, where it is viewable and copyable by GP+ members.
Does the following:
\enumerate{
\item Move teaching-materials/ content from the WD (project working directory on GP-Studio) to GalacticPolymath Shared Drive (Read-Only access)
\item Create shortcut to teaching-materials/ in place of the moved folder
\item Rename teaching-materials with a full "MediumTitle" i.e. more descriptive than typical working title of project
\item Update teach-it_*.gsheet DownloadLinks to reflect these file rearrangements
\item Make the following updates to front-matter:
\itemize{
\item GdriveTeachMatPath: 'GalacticPolymath/MediumTitle'
}
}
}
