% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_fm.R
\name{open_fm}
\alias{open_fm}
\alias{fm_open}
\title{open_fm}
\usage{
open_fm(WD = "?")

fm_open(WD = "?")
}
\arguments{
\item{WD}{the working directory; passed to \code{\link[=parse_wd]{parse_wd()}}; default= "?"}
}
\description{
open the front-matter.yml for a project
}
\section{Functions}{
\itemize{
\item \code{fm_open()}: alias for \code{\link[=open_fm]{open_fm()}}

}}
