% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_script.R
\name{open_script}
\alias{open_script}
\title{Opens scripts available for a given project}
\usage{
open_script(WD = "?", WD_git = NULL)
}
\arguments{
\item{WD}{working directory; if "?" or "s" supplied, will get key values for all projects in the GP-Studio drive. "??" or "l" will get data for "GP-LIVE";  default="s"}
}
\description{
Opens scripts available for a given project
}
