% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_unit_json.R
\name{open_unit_json}
\alias{open_unit_json}
\title{Open the lesson.json for a Unit}
\usage{
open_unit_json(WD = "?")
}
\arguments{
\item{WD}{working directory; default=getwd(); if "?" supplied, will invoke \code{\link[=pick_lesson]{pick_lesson()}}. The basename of this working directory will then be used to find a match in the gp-lessons git project folder by calling \code{\link[=get_wd_git]{get_wd_git()}}. It's a little roundabout, but is consistent with lookups centering on the Google Drive project working directory.}
}
\description{
Will find the compiled json and open in it in the system editor.
}
