% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_proj_folder.R
\name{open_wd}
\alias{open_wd}
\title{open_wd}
\usage{
open_wd(WD = "?", subpath = NULL)
}
\arguments{
\item{WD}{working directory; passed to \code{\link[=parse_wd]{parse_wd()}}.}

\item{subpath}{subpath to concat to WD path e.g.c("assets","data","teaching-materials"); default=NULL}
}
\description{
open_wd
}
\section{Functions}{
\itemize{
\item \code{open_wd()}: Alias for \link{open_proj_folder}

}}
