% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_matching.R
\name{overwrite_matching}
\alias{overwrite_matching}
\title{overwrite_matching}
\usage{
overwrite_matching(list1, list2)
}
\arguments{
\item{list1}{The source list (i.e. with newer data)}

\item{list2}{The target list (i.e. where new data will replace old data)}
}
\value{
List2 with matching elements of List1 overwritten over (outdated) List 2 values
}
\description{
For 2 lists, take matching values from List 1 and overwrite them on List 2
}
\details{
The lists don't have to have the same dimensions. Output will look like list2 with matching elements from list1 subbed in
}
