% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_locale.R
\name{parse_locale}
\alias{parse_locale}
\title{parse_locale}
\usage{
parse_locale(current_data)
}
\arguments{
\item{current_data}{the reconciled data including yaml and input from the shiny app environment; at a minimum you can pass list(Country="Whatever Country")}
}
\value{
current data with locale added
}
\description{
Takes the language and country chosen in the \code{\link[=editor]{editor()}} and creates a locale with a 3 digit language code and (if applicable), 2 digit country code. Only meant to run in the \code{\link[=compile_unit]{compile_unit()}} function.
}
