% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_wd.R
\name{parse_wd}
\alias{parse_wd}
\title{parse_wd}
\usage{
parse_wd(str = NULL)
}
\arguments{
\item{str}{is a string supplied as the WD parameter in a galacticPubs function. Often "?"}
}
\value{
if str is not "?" or "??", it will return str. Otherwise, returns the result of \code{\link[=pick_lesson]{pick_lesson()}}
}
\description{
Helper function to handle Working Directory strings. If "?" supplied, this invokes \link{pick_lesson("s")} to choose from among lessons in the 'GP-Studio' shared drive. If "??" supplied, this invokes \link{pick_lesson("l")} to pick a path to lessons on the 'GP-LIVE' drive.
}
