% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGPmarkdown.R
\encoding{UTF-8}
\name{parseGPmarkdown}
\alias{parseGPmarkdown}
\title{parseGPmarkdown}
\usage{
parseGPmarkdown(x, WD = NULL, mlinks = NULL, force_lookup = FALSE)
}
\arguments{
\item{x}{a text string to parse}

\item{WD}{working directory; default=NULL}

\item{mlinks}{multimedia info read in from the teach-it.gsheet!Titles tab}

\item{force_lookup}{logical; do you want to force looking up \code{teach-it.gsheet*!multimedia}? default=FALSE}
}
\value{
formatted replacement text
}
\description{
Functions to replace custom expressions with appropriate links or text. Custom tags include "\{vid 1\}", "\{vid 2\}" or "\{item 1\}", "\{item 2\}" more generally. Both have the same effect, but vid is clearer when reading and enforces video type, while item is more flexible. Symbols for media types: 'video'= ▶, 'pdf' or default= ➚. Special case for media titles with keyword "Cards"= ♧
}
\seealso{
Other markdown:
\code{\link[=expand_md_links]{expand_md_links()}},
\code{\link[=fixAnchorLinks]{fixAnchorLinks()}}
}
\concept{markdown}
