% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_input.R
\name{prep_input}
\alias{prep_input}
\title{Prep Input for Shiny App \code{\link[=editor]{editor()}}}
\usage{
prep_input(input, yaml_path = NULL, existing_current_data = NULL, WD = getwd())
}
\arguments{
\item{input}{shiny input variable}

\item{yaml_path}{relative path to front-matter.yml; default=NULL; preferred way is to provide WD}

\item{existing_current_data}{optional; reactive 'vals' variable containing certain environmental variables that don't have direct user inputs, but are included in front-matter.yml output. Pass in existing_current_data if you want to preserve reactive values (from being overwritten) that are not found in input or yaml.}

\item{WD}{working directory}
}
\value{
a list that contains saved_data and current_data, allowing to check for unsaved changes
}
\description{
Internal function not intended for users. Basically reconciles saved data, input data from shiny, and reactive values not included in input. This function is used to tell the front end whether there are unsaved changes.
}
