% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_files.R
\name{rename_files}
\alias{rename_files}
\title{Rename files based on a pattern}
\usage{
rename_files(pattern, replacement, dir_path, ignore.case = FALSE, perl = TRUE)
}
\arguments{
\item{pattern}{a regular expression to be passed to \code{\link[=gsub]{gsub()}}; e.g. "Short[ ]?Title" Will grab 'ShortTitle' and 'Short Title' prefixes and replace them with the replacement string

[ ]: R:\%20}

\item{replacement}{the string passed to \code{\link[=gsub]{gsub()}} that will replace the selected pattern in your file names}

\item{dir_path}{the folder or containing all the files you want to be changed.}

\item{ignore.case}{if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching. default= FALSE}

\item{perl}{logical. Should Perl-compatible regexps be used? default=TRUE}
}
\description{
Searches a directory recursively and replaces files by passing arguments to \code{\link[=gsub]{gsub()}}
}
