% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_google_analytics.R
\name{report_google_analytics}
\alias{report_google_analytics}
\title{Make a markdown report for Google Analytics for GP's website}
\usage{
report_google_analytics(
  since = NULL,
  start_date = NULL,
  end_date = NULL,
  paths = NULL,
  property_id = NULL
)
}
\arguments{
\item{since}{integer; instead of putting start/end date, put number of days prior to today}

\item{start_date}{start date in "2020/01/01" format; if left blank will be today()-90}

\item{end_date}{end date in "2020/01/01" format; if left blank will be today()}

\item{paths}{URLs}

\item{property_id}{an optional parameter passed to \code{\link[=get_page_metrics]{get_page_metrics()}}}
}
\description{
Make a markdown report for Google Analytics for GP's website
}
