% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_gp_productivity.R
\name{report_gp_productivity}
\alias{report_gp_productivity}
\title{Make a report about GP's lesson and other outputs}
\usage{
report_gp_productivity()
}
\value{
A report about GP's lesson and other outputs
}
\description{
Like it says on the tin
}
