% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_user_stats.R
\name{report_user_stats}
\alias{report_user_stats}
\title{Report information about GP users}
\usage{
report_user_stats(
  verbosity = 1,
  start_date = Sys.Date() - 90,
  end_date = Sys.Date(),
  view_data = TRUE,
  shape = "new",
  dev = FALSE
)
}
\arguments{
\item{verbosity}{passed to \code{\link[httr2:req_perform]{httr2::req_perform()}}; default=1}

\item{start_date}{Date; start date for report; default=Sys.Date()-30}

\item{end_date}{Date; end date for report; default=Sys.Date()}

\item{view_data}{logical; do you want to open retrieved user data with \code{\link[=View]{View()}}? Default=TRUE}

\item{shape}{how do you want the data element to look? options = c("new","engaged"). Default "new" sorts by latest added users, and includes all. "engaged" Orders by totalSignIns and subsets only isTeacher==TRUE}

\item{dev}{logical; if FALSE (default), gets catalog from the production gp-catalog. Otherwise, from the dev catalog.}
}
\value{
Summary tibble
}
\description{
Provide basic information about users
}
