% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmNArows.R
\name{rmNArows}
\alias{rmNArows}
\title{rmNArows}
\usage{
rmNArows(d)
}
\arguments{
\item{d}{the data frame you want to filter}
}
\value{
filtered x data frame
}
\description{
Get rid of rows from a data frame that are entirely NA
}
