% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lesson_scripts.R
\name{run_lesson_scripts}
\alias{run_lesson_scripts}
\title{run_lesson_scripts}
\usage{
run_lesson_scripts(scripts, skip = NULL, WD = getwd())
}
\arguments{
\item{scripts}{a vector of scripts (no path needed, just basenames)}

\item{skip}{filename of a script to skip}

\item{WD}{is working directory of the project (useful to supply for shiny app, which has diff. working environment)}
}
\description{
Run all scripts in the scripts/ subfolder in the lesson directory. Scripts must not have any interdependencies, as a general design rule.
}
