% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_read_yaml.R
\name{safe_read_yaml}
\alias{safe_read_yaml}
\title{safe_read_yaml}
\usage{
safe_read_yaml(yaml_path, WD = NULL, eval.expr = TRUE, checkWD = TRUE)
}
\arguments{
\item{yaml_path}{full path to the front-matter.yml file (usually in format '/users/lotsaSubfolders/WorkingDirectory/meta/front-matter.yml')}

\item{WD}{alternate way to specify location. e.g. use "?" to invoke \code{\link[=pick_lesson]{pick_lesson()}}}

\item{eval.expr}{boolean; do you want to evaluate expression in yaml prepended with '!expr '? Default=TRUE}

\item{checkWD}{do you want to run \code{\link[=check_wd]{check_wd()}} which will produce warning messages if working directory assumptions aren't met?; default= TRUE}
}
\description{
Wrapper for \link[yaml:read_yaml]{read_yaml} that simplifies all null and missing data to NA. Creates front-matter.yml if not found
}
