% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_json.R
\name{save_json}
\alias{save_json}
\title{save_json}
\usage{
save_json(
  out,
  filename,
  pretty = TRUE,
  auto_unbox = TRUE,
  na = "null",
  null = "null",
  ...
)
}
\arguments{
\item{out}{a data frame to be output (passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}})}

\item{filename}{output filename (including path)}

\item{pretty}{logical; prettify output or make an unreadable, ugly json? default=TRUE}

\item{auto_unbox}{logical param from \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}; default=TRUE}

\item{na}{how to print NA values; default= "null"}

\item{null}{how to print null values; default= "null"}

\item{...}{other arguments to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}}
}
\description{
A shallow wrapper for \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}
}
