% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_drive_local_credentials.R
\name{set_drive_local_credentials}
\alias{set_drive_local_credentials}
\title{Set up Google Drive local user credentials (and path)}
\usage{
set_drive_local_credentials()
}
\value{
table of galacticPubs-related environmental variables, their values, and whether they were successfully set
}
\description{
Sets environmental variables used to locate the Edu/lessons folder that is emulated by \href{https://www.google.com/drive/download}{Google Drive for Desktop}
}
