% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_gpPal.R
\name{show_gpPal}
\alias{show_gpPal}
\title{show_gpPal}
\usage{
show_gpPal(pal = 1, subset = NULL, ...)
}
\arguments{
\item{pal}{which palette? 1=6 color basic, 2= extended, 3= subject palette}

\item{subset}{the first n colors to show or x1:x2 or c(1,3,5) for a specific subset pal is 1 by default}

\item{...}{additional parameters for \code{\link[colorspace]{swatchplot}}}
}
\description{
Show the GP palette
}
