% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all.R
\name{show_all}
\alias{show_all}
\title{Show all rows in a tibble}
\usage{
show_all(df)
}
\arguments{
\item{df}{a dataframe or tibble}
}
\description{
Takes a data frame or tibble input and outputs a tibble that shows all rows.
}
