% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_capitalize_first.R
\name{string_capitalize_first}
\alias{string_capitalize_first}
\title{Capitalize first letter of a string}
\usage{
string_capitalize_first(x)
}
\arguments{
\item{x}{text string}
}
\description{
Does what it says
}
\seealso{
Other simple text string manipulation functions:
\code{\link[=paste_valid]{paste_valid()}},
\code{\link[=string_parseCamel]{string_parseCamel()}}
}
\concept{simple text string manipulation functions}
