% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_chunk.R
\name{time_chunk}
\alias{time_chunk}
\title{Make a "Time Chunking" Graphic}
\usage{
time_chunk(WD = "?")
}
\arguments{
\item{WD}{is working directory of the project; default= getwd()}
}
\description{
Creates a special graph showing what part of the allotted time period is being referred to for a set of steps in a lesson procedure.
}
