% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeChunk.R
\name{timeChunk}
\alias{timeChunk}
\title{timeChunk}
\usage{
timeChunk(
  totTime,
  segLengths,
  destFolder = "assets/timeChunk/",
  prefix,
  Color = "#2c83c3",
  fadedColor = "#1A4E75",
  lineWidth = 7
)
}
\arguments{
\item{totTime}{the total estimated time of}

\item{segLengths}{lengths (in min) of procedural steps (format= c(x1,x2,x3))}

\item{destFolder}{where to save the output graph}

\item{prefix}{add a prefix to graph names?}

\item{Color}{color of current step bar graph fill color}

\item{fadedColor}{color of prior, completed steps bar graph fill color}

\item{lineWidth}{width of each bar in the graph}
}
\description{
create visuals to show how lesson breakdown into chunks
}
