% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_sans_na.R
\name{unique_sans_na}
\alias{unique_sans_na}
\title{unique_sans_na}
\usage{
unique_sans_na(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector of the unique entries
}
\description{
Supply a vector, return a vector with just unique values, excluding NA
}
