% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_init.R
\name{unit_init}
\alias{unit_init}
\alias{init_unit}
\title{unit_init}
\usage{
unit_init(recover = FALSE, WD = "?")

init_unit(recover = FALSE, WD = "?")
}
\arguments{
\item{recover}{logical; do you want to try to finish initializing a unit that didn't complete due to some error? default=FALSE}

\item{WD}{working directory, passed to \code{\link[=parse_wd]{parse_wd()}}; default="?"; NOTE: only used if recover=T}
}
\description{
Set up a new Galactic Polymath lesson or mini-unit project folder. Create folder structure in 'GP-Studio'/Edu/Lessons Google Shared Drive and add meta/files
}
\details{
Will fail if ShortTitle duplicates an existing name found in GP-Studio or GP-Live
}
\section{Functions}{
\itemize{
\item \code{init_unit()}: alias for init_unit

}}
