% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_init_helper.R
\name{unit_init_helper}
\alias{unit_init_helper}
\title{Run unit_init_helper Shiny app}
\usage{
unit_init_helper()
}
\description{
Lets you set the following values for your lesson:
\itemize{
\item ShortName
\item Language
\item Country
\item LsnCount
}
}
\details{
Selections are passed to the global environment and interpreted by \code{\link[=unit_init]{unit_init()}}
}
