% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_drive_links.R
\name{update_drive_links}
\alias{update_drive_links}
\title{update_drive_links}
\usage{
update_drive_links(
  WD = getwd(),
  rebuild = NULL,
  rm_missing = TRUE,
  clean = FALSE
)
}
\arguments{
\item{WD}{a local virtualized path to a lesson folder where Google Drive (Web) path will be extracted from front matter. Easiest is to pass WD from \code{\link[=pick_lesson]{pick_lesson()}}; must use \code{full_path=TRUE} with pick_lesson}

\item{rebuild}{if T, rebuild everything; overrides checks of last modified times before updating links and teach-it.gsheet; default= NULL}

\item{rm_missing}{logical; do you want to automatically remove records that are missing (and for which no studioLink was found during crawl of Google Drive folder?); default= TRUE; if FALSE, the link will be left blank}

\item{clean}{logical; do you want to ignore all info on the teach-it.gsheet and only import inferred info from crawled google drive project files? THIS WILL OVERWRITE MANUALLY ENTERED TITLES; default=FALSE}
}
\description{
Searches lesson's google drive folder to populate the 'DriveLinks' tab of the lesson's 'meta/teach-it.gsheet' file.
}
\details{
It's a hybrid lookup system that requires Google Drive for Desktop to be setup with permission to GP-Studio shared drive and a web connection. Steps:
\enumerate{
\item Find WD working directory corresponding to the chosen lesson
\item Access the cloud version of the project folder by finding the 'GdriveDirID' in meta/front-matter.yml.
\item Search through web folders, compiling information from folder and naming structure in 'teaching-materials' folder.
\item Get Gdrive links for downloads, classroom- and remote- lessons, and assessments.
\item Merge results with manually entered titles and such in 'meta/teach-it.gsheet DriveLinks tab'
\item Save 'meta/teach-it.gsheet'
}
}
\seealso{
Other Google Drive Functions: 
\code{\link{drive_contents}()},
\code{\link{drive_copy_missing_files}()},
\code{\link{drive_cp_dir}()},
\code{\link{drive_find_path}()},
\code{\link{drive_get_info}()},
\code{\link{drive_inSync}()},
\code{\link{drive_new_from_template}()},
\code{\link{drive_open}()},
\code{\link{init_lesson_meta}()}
}
\concept{Google Drive Functions}
