% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_drive_links_partHelper.R
\name{update_drive_links_partHelper}
\alias{update_drive_links_partHelper}
\title{update_drive_links_partHelper}
\usage{
update_drive_links_partHelper(dribble, set_grades = NULL, set_envir = NULL)
}
\arguments{
\item{dribble}{the dribble for a "Part" subfolder of a lesson}

\item{set_grades}{passed to\code{\link[=drive_get_info]{drive_get_info()}}}

\item{set_envir}{passed to \code{\link[=drive_get_info]{drive_get_info()}}}
}
\description{
Helper function for \code{\link[=update_drive_links]{update_drive_links()}} that extracts information for each part (if found). Only intended for internal use in update_drive_links. Might eventually have it rename parts according titles in meta/teach-it.gsheet
}
