% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_meta_spreadsheet.R
\name{upgrade_meta_spreadsheet}
\alias{upgrade_meta_spreadsheet}
\title{Upgrade a meta/spreadsheet for a lesson or mini-unit}
\usage{
upgrade_meta_spreadsheet(WD = "?", template = NULL, force_upgrade = FALSE)
}
\arguments{
\item{WD}{default= "?"; The working directory, passed to \code{\link[=parse_wd]{parse_wd()}}}

\item{template}{which template do you want to use; default=NULL; values will be passed to \code{\link[=drive_find_path]{drive_find_path()}}}

\item{force_upgrade}{boolean; do you want to override a version check to see if template matches spreadsheet version (on tab1 in cell E1); default= FALSE}
}
\value{
boolean; did all steps succeed?
}
\description{
Transfer worksheet data from teach-it_ProjectName.gsheet or standards_ProjectName.gsheet to the latest template format. This keeps the project up-to-date with the latest galacticPubs workflow.
}
