% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_timesheets.R
\name{upgrade_timesheets}
\alias{upgrade_timesheets}
\title{Upgrade timesheet google sheets}
\usage{
upgrade_timesheets(
  timesheet = NULL,
  force_upgrade = FALSE,
  transfer_sharing = TRUE,
  update_description = NULL,
  relink = FALSE
)
}
\arguments{
\item{timesheet}{URL for the timesheet you want to upgrade; if left blank, will prompt you to choose from a list of available timesheets}

\item{force_upgrade}{boolean; do you want to override a version check to see if template matches spreadsheet version (on tab1 in cell E1); default= FALSE}

\item{transfer_sharing}{boolean; do you want to transfer sharing permissions from the old file to the new file? Will also send an email notifying them of the new file. default= TRUE}

\item{update_description}{details of what's new about this updated version. Will be tacked onto email when the new file it shared with the same email addresses as the old file}

\item{relink}{boolean; do you want to check/reestablish the Google Sheet link on GP-Projects.gsheet? default=F}
}
\value{
boolean; did all steps succeed?
}
\description{
Transfer timesheet data to the latest galacticPubs template sheet
}
