% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_assets.R
\name{upload_assets}
\alias{upload_assets}
\alias{gcs_upload_assets}
\title{Synchronize media assets with the cloud}
\usage{
upload_assets(WD = "?", bucket = "gp-cloud", clear = FALSE, backup = FALSE)

gcs_upload_assets(WD = "?", bucket = "gp-cloud", clear = FALSE, backup = FALSE)
}
\arguments{
\item{WD}{WD working directory, passed to \code{\link[=parse_wd]{parse_wd()}}}

\item{bucket}{which bucket? default= "gp-cloud"}

\item{clear}{do you want to clear all existing files for this project before uploading? default=FALSE}

\item{backup}{passed to \code{\link[=update_fm]{update_fm()}}; do you want to back up front-matter.yml before updating it with new URLs? default=FALSE to prevent too many backups}
}
\value{
invisibly returns a Tibble with filenames and download links
}
\description{
Upload lesson assets to galacticPubs Google Cloud Storage bucket. It looks for the following patterns in the following folders:
\itemize{
\item SponsorLogo  | assets/_banners_logos_etc: anything with 'logo' and .png, .jpg, .webp suffix
\item UnitBanner | assets/_banners_logos_etc: ideally just named banner.png; should work with 'ShortTitle_banner.png' and exclude 'old_banner.png'
\item LessonTiles  | assets/_banners_logos_etc: must be format 'L1_tile.png'; other img formats also supported: .png, .jpeg, .webp
\item UnitCard   | assets/_banners_logos_etc: format 'card.png' or 'ShortTitle.png'
\item QR-code | assets/_banners_logos_etc: format 'ProjTitle_QR-code.png'
\item LearningChart| assets/_learning-plots: file must be called 'GP-Learning-Chart.png' (autogenerated from \code{\link[=learningChart]{learningChart()}})
\item LearningEpaulette | assets/_learning-plots: file must be called 'GP-Learning-Epaulette.png' (autogenerated from \code{\link[=learningEpaulette]{learningEpaulette()}})
\item LearningEpaulette_vert | assets/_learning-plots: file must be called 'GP-Learning-Epaulette_vert.png' (autogenerated from \code{\link[=learningEpaulette]{learningEpaulette()}})
\item SupportingMedia | assets/_other-media-to-publish: anything except help.txt found in this folder
}

Alias for upload_assets
}
\section{Functions}{
\itemize{
\item \code{gcs_upload_assets()}: alias

}}
\seealso{
Other google cloud storage:
\code{\link[=gcs_add]{gcs_add()}},
\code{\link[=gcs_contents]{gcs_contents()}},
\code{\link[=gcs_delete]{gcs_delete()}}
}
\concept{google cloud storage}
