% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zassign_lsn_stats.R
\name{zassign_lsn_stats}
\alias{zassign_lsn_stats}
\title{zassign_lsn_stats}
\usage{
zassign_lsn_stats(is_initialized, WD_git, fm, uinfo)
}
\arguments{
\item{is_initialized}{logical; have the titles been changed from boiler plate?}

\item{WD_git}{location of gp-lessons github repo. Default=NULL will get this for the current workspace with \code{\link[=get_wd_git]{get_wd_git()}}}

\item{fm}{front-matter from \code{\link[=get_fm]{get_fm()}}}

\item{uinfo}{the unit info (tab 1 of teach-it.gsheet)}
}
\description{
A helper function for \code{\link[=compile_teach_it]{compile_teach_it()}}. Interprets statuses from Tab 1 of the teach-it.gsheet for a project. Writes these to front-matter with \code{\link[=update_fm]{update_fm()}}.
}
