% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zget_drive_id.R
\name{zget_drive_id}
\alias{zget_drive_id}
\title{zget_drive_id}
\usage{
zget_drive_id(
  drive_path,
  drive_root = NULL,
  exact_match = TRUE,
  fm_key = NULL,
  missing_val = NA
)
}
\arguments{
\item{drive_path}{passed to \code{\link[=drive_find_path]{drive_find_path()}}; default=NULL}

\item{exact_match}{passed to \code{\link[=drive_find_path]{drive_find_path()}}; default=TRUE}

\item{fm_key}{name of the front-matter key you're trying to fill; used in error messaging and naming output}

\item{missing_val}{what to return if key not matched. default=NA}
}
\value{
an ID as a single character value, named with fm_key parameter
}
\description{
Helper function used by \code{\link[=update_fm]{update_fm()}} that attempts to resolve a Google Drive (for Web) path, given a string, and return the id for this path. If it's not found, it should return NULL.
}
\details{
Primary usage is to populate 'Gdrive*ID' keys in the front-matter.yml. These are used to go between the virtualized "local" paths of the Google Drive for Desktop and the web paths used in the Web-based Google Drive API needed to create Google Drive files.
}
