% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zget_procedure.R
\name{zget_procedure}
\alias{zget_procedure}
\title{zget_procedure}
\usage{
zget_procedure(proc, lext, uinfo, mlinks)
}
\arguments{
\item{proc}{the procedure file that has already been read in from the teach-it.gsheet, i.e. via \code{\link[=compile_teach_it]{compile_teach_it()}}}

\item{lext}{the individual lesson extension info read from the teach-it.gsheet}

\item{uinfo}{the unit info read in from the teach-it.gsheet}

\item{mlinks}{multimedia info read in from the teach-it.gsheet!Titles tab}
}
\value{
a list of: the procedure chunks and Lesson-Extension links for each lesson and a compiled vocab dataframe
}
\description{
A helper function for \code{\link[=compile_teach_it]{compile_teach_it()}}. Translate the 'Procedure' tab of the lesson's 'teach-it.gsheet' into JSON format.
}
