% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zget_rsrcs.R
\name{zget_rsrcs}
\alias{zget_rsrcs}
\title{zget_rsrcs()}
\usage{
zget_rsrcs(df, fm)
}
\arguments{
\item{df}{a dataframe (tibble)}

\item{fm}{front matter info from \code{\link[=get_fm]{get_fm()}}}
}
\description{
Internal helpers for \code{\link[=compile_teach_mat]{compile_teach_mat()}}. Used to pull out highly structured hierarchical JSON from the teach-it.gsheet
}
\details{
Not intended for use outside functions.
}
\seealso{
Other Internal helper functions: 
\code{\link{zYTembed}()},
\code{\link{zcatchLinkNA}()},
\code{\link{zget_grade_bands}()}
}
\concept{Internal helper functions}
