% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zrename_lessons.R
\name{zrename_lessons}
\alias{zrename_lessons}
\title{zrename_lessons}
\usage{
zrename_lessons(uinfo, tmID, prompt_rename = TRUE)
}
\arguments{
\item{uinfo}{pinfo object passed from compile_teach_it()}

\item{tmID}{Google Drive ID for the folder where we'll find teaching materials. Depends on PublicationStatus. If it's "Draft", teaching-materials/ is found on GP-Studio; if "Live", it'll be on GalacticPolymath/ and view only}

\item{prompt_rename}{logical, do you want to promput user about whether to rename lessons? default=T}
}
\value{
logical of success of renaming; NA if nothing to change or no valid pinfo
}
\description{
Helper function for \code{\link[=compile_teach_it]{compile_teach_it()}}. Renames lesson folders based on information found in the project teach_it_*.gsheet!Titles.
}
