% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zrename_parts.R
\name{zrename_parts}
\alias{zrename_parts}
\title{zrename_parts}
\usage{
zrename_parts(pinfo, WD)
}
\arguments{
\item{pinfo}{pinfo object passed from compile_teach_it()}

\item{WD}{working directory}
}
\value{
logical of success of renaming
}
\description{
Helper function for \code{\link[=compile_teach_it]{compile_teach_it()}}. Renames part folders based on information found in the project teach_it_*.gsheet!PartTitles.
}
