% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_publish.R
\name{batch_publish}
\alias{batch_publish}
\title{batch_publish}
\usage{
batch_publish(commit_msg = NULL, shortName, lessons_dir)
}
\arguments{
\item{commit_msg}{What do you want to say about this update? Default=NULL, i.e. "automated galacticPubs::publish()"}

\item{shortName}{The lesson short name(s) (i.e. the folder name(s)) for the lesson(s) you want to update and rebuild; use "all" if you want to rebuild all the lessons. If left blank, a picker will try to give you a list of optionsefx}

\item{lessons_dir}{path to the virtualized folder Edu/lessons, where all the lessons are found; default="/Volumes/GoogleDrive/My Drive/Edu/Lessons/assumptionsMatter_femalesSing_math"}
}
\description{
Publish one or more lessons that have been staged in the published/ folder of a lesson directory
}
\details{
This assumes that you have Google Drive for Desktop set up and have permissions to access the lesson files.
}
