% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_publish.R
\name{batch_publish}
\alias{batch_publish}
\title{batch_publish}
\usage{
batch_publish(
  commit_msg = NULL,
  gh_proj_name = NULL,
  try_harder = FALSE,
  lessons_dir = NULL
)
}
\arguments{
\item{commit_msg}{What do you want to say about this update? Default=NULL, i.e. "automated galacticPubs::publish()"}

\item{gh_proj_name}{The unique project title of this lesson which is prefixed on the lesson folder name and the GitHub project. Not necessarily the same as the ShortTitle used in naming lesson presentations and worksheets; probably more specific with underscores; If left off, will try to get this info from the GitHubURL if available in the front-matter.yml.}

\item{try_harder}{Do you want the function to retry if it fails? A bit experimental. Gets passed to \code{\link[=catch_err]{catch_err()}}; default=FALSE}

\item{lessons_dir}{path to the virtualized folder Edu/lessons, where all the lessons are found; default=NULL}
}
\description{
Publish one or more lessons that have been staged in the published/ folder of a lesson directory
}
\details{
This assumes that you have Google Drive for Desktop set up and have permissions to access the lesson files.
}
