% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_rebuild.R
\name{batch_rebuild}
\alias{batch_rebuild}
\title{batch_rebuild}
\usage{
batch_rebuild(
  shortName,
  lessons_dir,
  stage = TRUE,
  change_this = NULL,
  clean = FALSE
)
}
\arguments{
\item{shortName}{What is the lesson short name (i.e. the folder name) for the lesson you want to update and rebuild; use "all" if you want to rebuild all the lessons.}

\item{lessons_dir}{path to the virtualized folder Edu/lessons, where all the lessons are found; default="/Volumes/GoogleDrive/My Drive/Edu/Lessons/assumptionsMatter_femalesSing_math"}

\item{stage}{do you want to stage assets that are rebuilt in the published/ folder for that lesson (prep to be published)? default=TRUE}

\item{change_this}{A list of values to change in the front matter before rebuilding. Default=NULL. Example: list(Title="Stormy Misty's Foal") would change the title of the lesson to the name of a horsey novel. If shortName=="all", make sure you set this to something you want to change for everything.}

\item{clean}{Do you want to clean the meta/JSON folder and build everything from scratch? (Gets passed to \code{\link[=compile_lesson]{compile_lesson()}}). Default=FALSE}
}
\description{
Function to rebuild a lesson or lessons that is (are) in a virtualized Google Drive folder. It will also stage the lesson in the published folder if requested, but will not publish.
}
\details{
This assumes that you have Google Drive for Desktop set up and have permissions to access the lesson files.
}
