% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_rebuild.R
\name{batch_rebuild}
\alias{batch_rebuild}
\title{batch_rebuild}
\usage{
batch_rebuild(
  gh_proj_name,
  lessons_dir,
  stage = TRUE,
  change_this = NULL,
  clean = TRUE,
  complete_rebuild = FALSE
)
}
\arguments{
\item{gh_proj_name}{The unique project title of this lesson which is prefixed on the lesson folder name and the GitHub project. Not necessarily the same as the ShortTitle used in naming lesson presentations and worksheets; probably more specific with underscores; If left off, will try to get this info from the GitHubPath if available in the front-matter.yml.}

\item{lessons_dir}{path to the virtualized folder Edu/lessons, where all the lessons are found; default="/Volumes/GoogleDrive/My Drive/Edu/Lessons"}

\item{stage}{do you want to call \code{\link[=stageAssets]{stageAssets()}} to stage files in the published/ folder for the lesson (i.e. prep to be published)? default=TRUE}

\item{change_this}{A list of values to change in the front matter before rebuilding. Default=NULL. Example: list(Title="Stormy Misty's Foal") would change the title of the lesson to the name of a horsey novel. If gh_proj_name=="all", make sure you set this to something you want to change for everything.}

\item{clean}{Do you want to clean the meta/JSON folder and build everything from scratch? (Gets passed to \code{\link[=compile_lesson]{compile_lesson()}}). Default=TRUE}

\item{complete_rebuild}{Do you want to force rebuild everything (even if a particular item seems up to date?) default=FALSE (This par gets passed on as rebuild to \code{\link[=compile_lesson]{compile_lesson()}})}
}
\description{
Function to rebuild a lesson or lessons that is (are) in a virtualized Google Drive folder. It will also stage the lesson in the published folder if requested, but will not publish.
}
\details{
This assumes that you have Google Drive for Desktop set up and have permissions to access the lesson files.
}
