% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCompile.R
\name{batchCompile}
\alias{batchCompile}
\title{batchCompile}
\usage{
batchCompile(
  current_data,
  choices = c("Front Matter"),
  destFolder,
  outputFileName = "LESSON.json",
  WD = getwd(),
  img_loc,
  clean = FALSE,
  rebuild = NULL
)
}
\arguments{
\item{current_data}{the reconciled data including yaml and input from the shiny app environment; if current_data=NULL, read in front-matter.yml}

\item{choices}{one or more of the following: c("Front Matter","Standards Alignment","Teaching Materials","Procedure","Acknowledgements","Versions"); or "All"}

\item{destFolder}{where you want to save the folder; by default in the "meta/JSON/" folder}

\item{outputFileName}{output file name; default= "processedProcedure.json"}

\item{WD}{is working directory of the project (useful to supply for shiny app, which has diff. working environment)}

\item{img_loc}{where files are being stored (www folder)}

\item{clean}{delete all JSON files in meta/ and start over? default=FALSE}

\item{rebuild}{if T, rebuild everything; overrides RebuildAllMaterials in front-matter.yml; default= NULL}
}
\value{
a JSON is saved to meta/JSON/LESSON.json
}
\description{
Processes a subset of data chosen by user using the GP Shiny Editor
}
\details{
Combines functionality of compileProcedure, compileStandards, compileAcknowledgements, compileJSON, etc.
}
