% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCompile.R
\name{batchCompile}
\alias{batchCompile}
\title{batchCompile}
\usage{
batchCompile(
  input,
  choices = c("Front Matter"),
  destFolder = "meta/JSON/",
  outputFileName = "LESSON.json",
  WD = getwd()
)
}
\arguments{
\item{input}{the input from the shiny app environment}

\item{choices}{one or more of the following: c("Front Matter","Standards Alignment","Teaching Materials","Procedure","Acknowledgements","Versions")}

\item{destFolder}{where you want to save the folder; by default in the "meta/JSON/" folder}

\item{outputFileName}{output file name; default= "processedProcedure.json"}

\item{WD}{is working directory of the project (useful to supply for shiny app, which has diff. working environment)}
}
\value{
a JSON is saved to meta/JSON/LESSON.json
}
\description{
Processes a subset of data chosen by user using the GP Shiny Editor
}
\details{
Combines functionality of compileProcedure, compileStandards, compileAcknowledgements, compileJSON, etc.
}
