% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileAlignment.R
\name{compileAlignment}
\alias{compileAlignment}
\title{compileAlignment}
\usage{
compileAlignment(
  lessonAlignmentMatrix,
  grades = "all",
  fileName = "processedStandards.json"
)
}
\arguments{
\item{lessonAlignmentMatrix}{file location of the lesson alignment matrix XLSX worksheet}

\item{grades}{grade band on alignment matrix worksheet options= "5-6", "7-8", "9-12" or "all"}

\item{fileName}{output filename (including path if you don't want it in the working directory)}
}
\description{
Compile alignment info from a lesson standards matrix worksheet
}
