% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileAlignment.R
\name{compileAlignment}
\alias{compileAlignment}
\title{compileAlignment}
\usage{
compileAlignment(
  lessonAlignmentMatrix,
  grades = "all",
  destFolder = "standards/",
  fileName = "processedStandards.json"
)
}
\arguments{
\item{lessonAlignmentMatrix}{file location of the lesson alignment matrix XLSX worksheet}

\item{grades}{grade band on alignment matrix worksheet options= "5-6", "7-8", "9-12" or "all"}

\item{destFolder}{where you want to save the folder; by default in the "standards/" folder, 1 level up from the working directory}

\item{fileName}{output file name; default= "processedStandards.json"}
}
\value{
tibble of the compiled standards data; a JSON is saved to standards/processedStandards.json
}
\description{
Compile alignment info from a lesson standards matrix worksheet
}
