% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileJSON.R
\name{compileJSON}
\alias{compileJSON}
\title{compileJSON}
\usage{
compileJSON(outputFileName = "LESSON.json", WD = getwd())
}
\arguments{
\item{outputFileName}{output file name; default= "processedProcedure.json"}

\item{WD}{is working directory of the project (useful to supply for shiny app, which has diff. working environment)}

\item{destFolder}{where you want to save the folder; by default in the "meta/JSON/" folder}
}
\value{
tibble of the compiled standards data; a JSON is saved to meta/JSON/processedProcedure.json
}
\description{
Combine all JSON components into 1 compiledLesson.JSON
}
