% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileJSON.R
\name{compileJSON}
\alias{compileJSON}
\title{compileJSON}
\usage{
compileJSON(destFolder = "meta/JSON/", outputFileName = "LESSON.json")
}
\arguments{
\item{destFolder}{where you want to save the folder; by default in the "meta/JSON/" folder}

\item{outputFileName}{output file name; default= "processedProcedure.json"}
}
\value{
tibble of the compiled standards data; a JSON is saved to meta/JSON/processedProcedure.json
}
\description{
Combine all JSON components into 1 compiledLesson.JSON
}
