% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileTeachingMat.R
\name{compileTeachingMat}
\alias{compileTeachingMat}
\title{compileTeachingMat}
\usage{
compileTeachingMat(
  linksFile = "meta/teaching-materials.xlsx",
  procedureFile = "meta/procedure.xlsx",
  destFolder = "meta/JSON/",
  outputFileName = "teachingMaterials.json"
)
}
\arguments{
\item{linksFile}{file location of the lesson teaching-materials XLSX worksheet.}

\item{procedureFile}{file location of the lesson procedure XLSX worksheet}

\item{destFolder}{where you want to save the folder; by default in the "meta/JSON/" folder}

\item{outputFileName}{output file name; default= "processedTeachingMaterials.json"}
}
\value{
tibble of the compiled standards data; a JSON is saved to meta/JSON/processedResources.json
}
\description{
Compile Teaching Materials data from teaching-materials.xlsx
}
