% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileVersions.R
\name{compileVersions}
\alias{compileVersions}
\title{compileVersions}
\usage{
compileVersions(
  inputFileName = "meta/version-info.xlsx",
  destFolder = "meta/JSON/",
  outputFileName = "versions.json"
)
}
\arguments{
\item{inputFileName}{file location of the lesson alignment matrix XLSX worksheet; include full path if not in working directory}

\item{destFolder}{where you want to save the folder; by default in the "JSON/" folder}

\item{outputFileName}{output file name; default= "acknowledgments.json"}
}
\value{
acknowledgment list object; a JSON is saved to standards/acknowledgments.json
}
\description{
Compile version info from an XLSX to a JSON file
}
