% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_changes.R
\name{discard_changes}
\alias{discard_changes}
\title{Discard local changes to git-watched files}
\usage{
discard_changes(gh_proj_name, lessons_dir)
}
\arguments{
\item{gh_proj_name}{The unique project title of this lesson as it is named on \url{https://github.com/galacticpolymath}. Not \emph{necessarily} the same as the ShortTitle used in naming lesson presentations and worksheets; probably more specific with underscores. If left off, will try to get this info from the GitHubPath if available in the front-matter.yml.}

\item{lessons_dir}{path to the virtualized folder Edu/lessons, where all the lessons are found; default="/Volumes/GoogleDrive/My Drive/Edu/Lessons"}
}
\description{
This will revert local changes for a lesson project folder. Does not \emph{currently} include files that are not watched (e.g. in assets/ or assembled-lesson-materials/ folders)
}
