% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_contents.R
\name{drive_contents}
\alias{drive_contents}
\title{drive_contents}
\usage{
drive_contents(drib)
}
\arguments{
\item{drib}{a Google drive dribble of a folder (usually the output of \code{\link[=drive_find_path]{drive_find_path()}})}
}
\value{
a dribble showing the contents of the desired folder
}
\description{
Show the contents of a Google drive folder, given a dribble
}
\details{
A shallow wrapper for \code{\link[googledrive:drive_ls]{googledrive::drive_ls()}} that uses just the id of a dribble object and can be piped from \code{\link[=drive_find_path]{drive_find_path()}}
}
\seealso{
Other Google Drive Functions: 
\code{\link{drive_copy_missing_files}()},
\code{\link{drive_cp_dir}()},
\code{\link{drive_find_path}()},
\code{\link{drive_open}()}
}
\concept{Google Drive Functions}
