% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_cp_dir.R
\name{drive_cp_dir}
\alias{drive_cp_dir}
\title{Clone a Google Drive Folder}
\usage{
drive_cp_dir(source_folder, dest_folder)
}
\description{
Copies all files (Google Workspace or otherwise) from one destination to another. This is a workaround for the absence of this functionality in the GDrive API. It's accomplished by recursive calls to \code{\link[googledrive:drive_cp]{googledrive::drive_cp()}}.
}
