% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_find_path.R
\name{drive_find_path}
\alias{drive_find_path}
\title{Find folder or file along Google Drive path}
\usage{
drive_find_path(drive_path)
}
\arguments{
\item{drive_path}{in the form "DRIVE/directory/subdirectory". DRIVE can be "~" or "my drive" to refer to your private google drive; it can also be the name of a shared drive (e.g. "GP-Misc"). Generally, case SeNsItIvE.}
}
\description{
Simulates Finder/File Explorer functionality by repeated calls to \code{\link[googledrive:drive_find]{googledrive::drive_find()}}.
}
\seealso{
Other Google Drive Functions: 
\code{\link{drive_contents}()},
\code{\link{drive_copy_missing_files}()},
\code{\link{drive_cp_dir}()},
\code{\link{drive_open}()}
}
\concept{Google Drive Functions}
