% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_find_path.R
\name{drive_find_path}
\alias{drive_find_path}
\title{Find folder or file along Google Drive path}
\usage{
drive_find_path(drive_path, WD = NULL, drive_root = NULL)
}
\arguments{
\item{drive_path}{If you provide a dribble, it gets passed right back out. Otherwise, you can give a character string in the form "DRIVE/directory/subdirectory".
\itemize{
\item DRIVE can be "~" or "my drive" to refer to your private google drive
\item Path can also be the name of a shared drive (e.g. "GP-Misc")
\item Also supports relative paths (e.g. "../meta"), as long as \strong{WD} or \strong{drive_root}. If \strong{WD} is supplied (this is a full "local" Google Drive for Desktop path piped from \code{\link[=pick_lesson]{pick_lesson()}}). If \strong{drive_root} supplied, it will use that as the parent for getting relative paths.
\item Relative pathing can be significantly faster because each hierarchical call to GDrive API to resolve a folder costs about a second.
\item You can also pass an ID as a text string if you know that already
\item Generally, case SeNsItIvE...but results may vary
}}

\item{WD}{\itemize{
\item a \strong{local} virtualized path to a lesson folder where Google Drive (Web) path will be extracted from front matter. Easiest is to pass "?" whcih will invoke \code{\link[=pick_lesson]{pick_lesson()}}; must use \code{full_path=TRUE} with pick_lesson
\item will be ignored unless relative path provided ("../folder1"), where \strong{WD} will be substituted for ".."
}}

\item{drive_root}{A Google drive path reference (not a local or virtualized path). Will be ignored unless relative path provided ("../folder1"), where drive_root will be substituted for "..". Can be:
\enumerate{
\item a dribble or
\item a Googledrive ID (as a string)
\item drive_root is passed to \code{\link[googledrive:drive_get]{googledrive::drive_get()}}
}}
}
\description{
Simulates Finder/File Explorer functionality by repeated calls to \code{\link[googledrive:drive_find]{googledrive::drive_find()}}. Allows relative paths via hybrid navigation of local virtualized Google Drive for Desktop paths by going to a local lesson directory, pulling up the front-matter.yml and finding the Google ID in order to access the relative path you requested through the Drive API.
}
\examples{
\dontrun{
#ABSOLUTE PATHS
#path to a drive file on your personal Google Drive
drive_find_path("~/folder_in_my_personal_drive/filename")
#path to a network drive folder
(p <- drive_find_path("GP-Studio/Edu/Lessons/geneticrescue_sci"))
#show contents of that drive folder
p \%>\% drive_contents()

#RELATIVE PATH to a particular lesson subfolder
#only works if you have Google Drive for Desktop set up with permissions to GP-Studio
drive_find_path("../assets",pick_lesson()) \%>\% drive_contents
}

}
\seealso{
Other Google Drive Functions: 
\code{\link{drive_contents}()},
\code{\link{drive_copy_missing_files}()},
\code{\link{drive_cp_dir}()},
\code{\link{drive_get_info}()},
\code{\link{drive_inSync}()},
\code{\link{drive_new_from_template}()},
\code{\link{drive_open}()},
\code{\link{init_lesson_meta}()},
\code{\link{update_drive_links}()}
}
\concept{Google Drive Functions}
