% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_find_path.R
\name{drive_find_path}
\alias{drive_find_path}
\title{Find folder or file along Google Drive path}
\usage{
drive_find_path(drive_path)
}
\arguments{
\item{drive_path}{in the form "directory/subdirectory", where "My Drive" is understood to be at the root above "directory". Will match case InSeNsItIvElY. If you want to place an item on the bare "My Drive" folder, set drive_path="~" or "root".}
}
\description{
Simulates Finder/File Explorer functionality by repeated calls to \code{\link[googledrive:drive_find]{googledrive::drive_find()}}.
}
\seealso{
Other Google Drive Functions: 
\code{\link{drive_cp_dir}()}
}
\concept{Google Drive Functions}
