% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_get_info.R
\name{drive_get_info}
\alias{drive_get_info}
\title{drive_get_info}
\usage{
drive_get_info(dribble, set_envir = NULL, set_grades = NULL, validate = FALSE)
}
\arguments{
\item{dribble}{a one row dribble input (e.g. piped from \code{\link[googledrive:drive_get]{googledrive::drive_get()}}) for a single file}

\item{set_envir}{set envir output manually (not from file name); partial string matching of options "classroom" and "remote"}

\item{set_grades}{set grade bands manually (not from file name); passed as a string (e.g. "5-9")}

\item{validate}{logical; do you want to throw an error if any of the following are missing? default= FALSE
\itemize{
\item shortTitle
\item title
\item part
\item grades
\item itemType
\item fileType
}}
}
\value{
a tibble with a row corresponding to each row in dribble input, with title and other information extracted from filename
}
\description{
Extract galacticPubs-relevant information from a GP lesson file, with an expected format of \emph{ShortTitle_P1_G5-9_wksht (TEACHER)}
}
\details{
Also extracts the modification datetime and drive link. 'title' and 'description' are also included to match DriveLinks fields, though they cannot be populated automatically.
}
\examples{
  drive_find_path("GP-Workshop/Edu/Lessons/assumptionsMatter_femalesSing_math/teaching-materials/classroom/classroom_5-6/handouts/Females Sing_P1_G5-6 wksht (STUDENT)_classroom.gdoc")

}
\seealso{
Other Google Drive Functions: 
\code{\link{drive_contents}()},
\code{\link{drive_copy_missing_files}()},
\code{\link{drive_cp_dir}()},
\code{\link{drive_find_path}()},
\code{\link{drive_inSync}()},
\code{\link{drive_new_from_template}()},
\code{\link{drive_open}()},
\code{\link{init_lesson_meta}()},
\code{\link{update_drive_links}()}
}
\concept{Google Drive Functions}
