% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_get_info.R
\name{drive_get_info}
\alias{drive_get_info}
\title{drive_get_info}
\usage{
drive_get_info(
  dribble,
  set_lsn = NULL,
  set_envir = NULL,
  set_grades = NULL,
  validate = FALSE,
  all_info = FALSE,
  include_shared_drive = FALSE
)
}
\arguments{
\item{dribble}{a dribble input (e.g. piped from \code{\link[googledrive:drive_get]{googledrive::drive_get()}}) for a single file}

\item{set_lsn}{set lesson manually (not from file name)}

\item{set_envir}{set envir output manually (not from file name); partial string matching of options "classroom", "remote", "assessments"}

\item{set_grades}{set grade bands manually (not from file name); passed as a string (e.g. "5-9")}

\item{validate}{logical; do you want to throw an error if any of the following are missing? default= FALSE}

\item{all_info}{logical; keep the full drive_resources list with all the dribble meta info for each row item? default=FALSE}

\item{include_shared_drive}{logical; do you want to include the name of the shared drive (if any) that the file is in? default=FALSE. If TRUE, will add a column: _sharedDrive (takes a little longer with another call to Google Drive API)
\itemize{
\item shortTitle
\item title
\item \verb{_lsn}
\item \verb{_grades}
\item \verb{_itemType}
\item \verb{_fileType}
}}
}
\value{
a tibble with a row corresponding to each row in dribble input, with title and other information extracted from filename. Possible outputs:
\itemize{
\item shortTitle: ShortTitle from filename (e.g. "femalesSing")
\item _lsn:
\item _grades: grade band (e.g. "5-9")
\item _itemType: guessed item type (e.g. "worksheet", "handout", "presentation", "card", "table", "overview", "assessment")
\item _fileType: Google Drive file type (e.g. "Google Docs", "PDF", "JPEG", "MP4 video", "Google Slides", "Google Sheets", "audio")
\item _envir: guessed environment (e.g. "classroom", "remote", or NA)
\item _SvT: guessed Student vs Teacher (e.g. "student", "teacher", or NA)
\item _link: Google Drive link (with edit? or view? removed)
\item modTime: modification datetime (POSIXct)
}
}
\description{
Extract galacticPubs-relevant information from a GP lesson file, with an expected format of \emph{ShortTitle_L1_G5-9_wksht (TEACHER)}
}
\details{
Also extracts the modification datetime and drive link. 'title' and 'description' are also included to match TeachMatLinks fields, though they cannot be populated automatically.
}
\examples{
\dontrun{
  drive_find_path("GP-Studio/Edu/Lessons/assumptionsMatter_femalesSing_math/teaching-materials/classroom/classroom_5-6/handouts/Females Sing_P1_G5-6 wksht (STUDENT)_classroom.gdoc")
  }

}
\seealso{
Other Google Drive Functions: 
\code{\link{drive_contents}()},
\code{\link{drive_copy_missing_files}()},
\code{\link{drive_cp_dir}()},
\code{\link{drive_find_path}()},
\code{\link{drive_inSync}()},
\code{\link{drive_new_from_template}()},
\code{\link{drive_open}()},
\code{\link{init_unit_meta}()},
\code{\link{update_teach_links}()}
}
\concept{Google Drive Functions}
