% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_new_from_template.R
\name{drive_new_from_template}
\alias{drive_new_from_template}
\title{drive_new_from_template()}
\usage{
drive_new_from_template(
  template_path = NULL,
  dest_path = NULL,
  new_name = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{template_path}{text string of a path to a template to pass to \code{\link[=drive_find_path]{drive_find_path()}}}

\item{dest_path}{text string of a path where you want to file to go}

\item{new_name}{text string of what you'd like to name your new file. default=NULL results in "Copy of FILENAME"; passed to \code{\link[googledrive:drive_cp]{googledrive::drive_cp()}}}

\item{overwrite}{default=FALSE; passed to \code{\link[googledrive:drive_cp]{googledrive::drive_cp()}}}
}
\description{
Create a new Galactic Polymath Google Drive file from a template
}
\seealso{
Other Google Drive Functions: 
\code{\link{drive_contents}()},
\code{\link{drive_copy_missing_files}()},
\code{\link{drive_cp_dir}()},
\code{\link{drive_find_path}()},
\code{\link{drive_open}()}
}
\concept{Google Drive Functions}
