% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_parse_name.R
\name{drive_parse_name}
\alias{drive_parse_name}
\title{drive_parse_name}
\usage{
drive_parse_name(dribble)
}
\arguments{
\item{dribble}{a one row dribble input (e.g. piped from \code{\link[googledrive:drive_get]{googledrive::drive_get()}}) for a single file}
}
\description{
Extract information from a GP lesson file, with an expected format of \emph{ShortTitle_P1_G5-9_wksht (TEACHER)}
}
\examples{
  drive_find_path("GP-Workshop/Edu/Lessons/assumptionsMatter_femalesSing_math/teaching-materials/classroom/classroom_5-6/handouts/Females Sing_P1_G5-6 wksht (STUDENT)_classroom.gdoc")
}
