% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor.R
\name{editor}
\alias{editor}
\title{editor}
\usage{
editor(pick = FALSE)
}
\arguments{
\item{pick}{logical; do you want to pick a lesson to work on? default=F (work from the current working directory)}
}
\value{
Outputs to meta/front-matter.yaml
}
\description{
Start a Shiny app to edit front matter for a lesson. Accepts markdown syntax. The code is found in inst/shiny/gp_editor
}
