% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor.R
\name{editor}
\alias{editor}
\title{editor}
\usage{
editor(WD = getwd(), system_browser = TRUE)
}
\arguments{
\item{WD}{a local virtualized path to a lesson folder where Google Drive (Web) path will be extracted from front matter. Easiest is to pass "?" which will invoke \code{\link[=pick_lesson]{pick_lesson()}}; default=getwd()}

\item{system_browser}{logical
\itemize{
\item TRUE (default): open in system default web browser
\item FALSE: open in Rstudio browser pane
}}
}
\value{
Outputs to meta/front-matter.yaml
}
\description{
Start a Shiny app to edit front matter for a lesson. Accepts markdown syntax. The code is found in inst/shiny/gp_editor
}
