% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcs_contents.R
\name{gcs_contents}
\alias{gcs_contents}
\title{Get contents of GP Google Cloud Storage bucket}
\usage{
gcs_contents(
  WD = "?",
  cloud_dir = NULL,
  bucket = "gp-cloud",
  detail = "summary",
  pattern = NULL,
  show_all = FALSE,
  ...
)
}
\arguments{
\item{WD}{WD working directory, passed to \code{\link[=parse_wd]{parse_wd()}}. Default="?" will pick unit. Supplying WD will give you the subset for a given unit.}

\item{cloud_dir}{default=NULL; which directory in the chosen bucket do you want to list? Options= "css","icons","lessons","logos","people"}

\item{bucket}{default="GP-Studio"; which bucket do you want to pick?}

\item{detail}{the detail parameter for \code{\link[googleCloudStorageR:gcs_list_objects]{googleCloudStorageR::gcs_list_objects()}}; options= c("summary", "more", "full"); default="more"}

\item{pattern}{default=NULL; pattern for filtering rows}

\item{show_all}{default=FALSE; if TRUE, will show all rows (may take a while)}

\item{...}{other parameters passed to \code{\link[googleCloudStorageR:gcs_list_objects]{googleCloudStorageR::gcs_list_objects()}}}
}
\value{
a Tibble with success, filenames and download links
}
\description{
A shallow wrapper for \code{\link[googleCloudStorageR:gcs_list_objects]{googleCloudStorageR::gcs_list_objects()}}. If WD supplied, it will only show objects for that project.
}
\seealso{
Other google cloud storage: 
\code{\link{gcs_add}()},
\code{\link{gcs_delete}()},
\code{\link{upload_assets}()}
}
\concept{google cloud storage}
