% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcs_contents.R
\name{gcs_contents}
\alias{gcs_contents}
\title{Get contents of GP Google Cloud Storage bucket}
\usage{
gcs_contents(WD = NULL, bucket = "gp-cloud", detail = "summary", ...)
}
\arguments{
\item{WD}{WD working directory, passed to \code{\link[=parse_wd]{parse_wd()}}. Default=NULL will show all bucket objects. Supplying WD will give you the subset for a given unit.}

\item{detail}{the detail parameter for \code{\link[googleCloudStorageR:gcs_list_objects]{googleCloudStorageR::gcs_list_objects()}}; options= c("summary", "more", "full"); default="more"}

\item{...}{other parameters passed to \code{\link[googleCloudStorageR:gcs_list_objects]{googleCloudStorageR::gcs_list_objects()}}}
}
\value{
a Tibble with success, filenames and download links
}
\description{
A shallow wrapper for \code{\link[googleCloudStorageR:gcs_list_objects]{googleCloudStorageR::gcs_list_objects()}}. If WD supplied, it will only show objects for that project.
}
\seealso{
Other google cloud storage: 
\code{\link{gcs_add}()},
\code{\link{gcs_delete}()},
\code{\link{upload_assets}()}
}
\concept{google cloud storage}
