% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcs_delete.R
\name{gcs_delete}
\alias{gcs_delete}
\title{Helper for \code{\link[=upload_assets]{upload_assets()}}: Delete a file from Google Cloud Storage bucket if not found locally}
\usage{
gcs_delete(cloud_path, bucket = "gp-cloud")
}
\arguments{
\item{cloud_path}{expects a vector of 1 or more cloud_paths in the form: 'lessons/PROJ/x.jpg'}

\item{bucket}{GCS bucket; default= "gp-cloud"}
}
\value{
tibble of successes
}
\description{
A shallow wrapper for \code{\link[googleCloudStorageR:gcs_delete_object]{googleCloudStorageR::gcs_delete_object()}} with GP error checking. Will validate:
\itemize{
\item file(s) exist
}
}
\seealso{
Other google cloud storage: 
\code{\link{gcs_add}()},
\code{\link{gcs_contents}()},
\code{\link{upload_assets}()}
}
\concept{google cloud storage}
