% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fm.R
\name{get_fm}
\alias{get_fm}
\title{Get front-matter values}
\usage{
get_fm(key = NULL, WD = getwd())
}
\arguments{
\item{key}{which entry (or entries) do you want to import? default=NULL will import everything}

\item{WD}{working directory; default=getwd()}
}
\description{
Imports meta/front-matter.yml value(s) to environment as a list.
}
\examples{
get_fm()
get_fm(key=c("Title","ShortTitle","locale"))
get_fm(key="Title",WD=pick_lesson())
}
