% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fm.R
\name{get_fm}
\alias{get_fm}
\title{Get front-matter values}
\usage{
get_fm(
  key = NULL,
  WD = "?",
  WD_git = NULL,
  yaml_path = NULL,
  checkWD = FALSE,
  auto_init = FALSE,
  check = NULL,
  always_list = FALSE,
  standardize_NA = TRUE,
  ...
)
}
\arguments{
\item{key}{which entry (or entries) do you want to import? default=NULL will import everything; Supports "starts with", case-insensitive matching for a single key if prefixed with '~'}

\item{WD}{working directory; default=getwd(); if "?" supplied, will invoke \code{\link[=pick_lesson]{pick_lesson()}}. The basename of this working directory will then be used to find a match in the gp-lessons git project folder by calling \code{\link[=get_wd_git]{get_wd_git()}}. It's a little roundabout, but is consistent with lookups centering on the Google Drive project working directory.}

\item{WD_git}{default=NULL. If you already know the path to the gp-lessons folder, this is more efficient.}

\item{yaml_path}{default=NULL. An alternate way to just provide a full path to read in a yml file.}

\item{checkWD}{passed to \code{\link[=safe_read_yaml]{safe_read_yaml()}}; default=FALSE; set to FALSE to suppress warnings if for example you're missing teach-it.gsheet or some other item expected to be in a lesson directory}

\item{auto_init}{logical; do you want to automatically create a front-matter.yml file if it's not found? Runs \code{\link[=init_fm]{init_fm()}}; default=FALSE}

\item{check}{string referring to a check function(x) to pass to \code{\link[checkmate:assert]{checkmate::assert()}}; e.g. check="checkmate::check_character(x,min.chars=10)" will throw an error if an output is not a string of at least 10 characters. default=NULL}

\item{always_list}{logical; do you want to always return a list? default=FALSE will unlist() the result if it seems to be a single item}

\item{standardize_NA}{logical; do you want all "",NULL,list(), etc. to be read in as NA using \code{\link[=is_empty]{is_empty()}}? passed to \code{\link[=safe_read_yaml]{safe_read_yaml()}}; default=TRUE}

\item{...}{additional args passed to check function}
}
\description{
Imports meta/front-matter.yml value(s) to environment as a list.
}
\details{
If you ask for only one key, output will be a vector, rather than a list
}
\examples{
get_fm()
get_fm(key=c("Title","ShortTitle","locale"))
get_fm(key="Title",WD=pick_lesson())
#partial matching for all front-matter entries starting with 'gdrive' (case insensitive)
get_fm("gdrive","?")
}
