% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_git_gp_lessons_path.R
\name{get_git_gp_lessons_path}
\alias{get_git_gp_lessons_path}
\title{get_git_gp_lessons_path}
\usage{
get_git_gp_lessons_path()
}
\value{
virtualized path to Google Shared Drives location
}
\description{
Gets a virtualized path to Google Shared Drives. Uses environmental variables set by \code{\link[=init_galacticPubs]{init_galacticPubs()}}
}
