% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gp_api_token.R
\name{get_gp_api_token}
\alias{get_gp_api_token}
\alias{gp_api_get_token}
\title{Get a JWT token for GP API}
\usage{
get_gp_api_token(trigger_oauth = FALSE, refresh = FALSE)

gp_api_get_token(trigger_oauth = FALSE, refresh = FALSE)
}
\arguments{
\item{trigger_oauth}{a temporary, ugly setting. If T, this will run oauth_flow_auth_code, which then hangs. Default=F; Rerun with False to get the token #smh}

\item{refresh}{do you want to re-authenticate? default=FALSE}
}
\value{
invisibly returns the token
}
\description{
Will get a token from the R environment if available and test it with an empty query to GP API. If missing or expired, will attempt to renew through Google Oauth flow.
}
\section{Functions}{
\itemize{
\item \code{gp_api_get_token()}: alias

}}
\seealso{
Other GP API: 
\code{\link{gp_api_query}()},
\code{\link{gp_api_unit_delete}()},
\code{\link{gp_api_unit_insert}()},
\code{\link{gp_api_unit_replace}()}
}
\concept{GP API}
