% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shared_drive_path.R
\name{get_shared_drive_path}
\alias{get_shared_drive_path}
\title{get_shared_drive_path}
\usage{
get_shared_drive_path()
}
\value{
virtualized path to Google Shared Drives location
}
\description{
Gets a virtualized path to Google Shared Drives. Uses environmental variables set by \code{\link[=set_drive_local_credentials]{set_drive_local_credentials()}}
}
