% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state.R
\name{get_state}
\alias{get_state}
\title{get_state}
\usage{
get_state(path, save_path = NULL)
}
\arguments{
\item{path}{path to the directory you want to collect info for (using local, i.e. Google Drive Virtual, path); can be a vector (e.g. c(path1, path2))}

\item{save_path}{if you want to save the output, write a local path; must be a full path from root to filename, including extension .RDS; default=NULL}
}
\value{
a tibble of recursive filenames, type, size, and modification_time for the path
}
\description{
Generate a list of info for a directory path to use to trigger an update if necessary.
}
