% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_rename_repo.R
\name{gh_rename_repo}
\alias{gh_rename_repo}
\title{Rename a Galactic Polymath lesson GitHub repository online}
\usage{
gh_rename_repo(new_proj_name, gh_proj_name, lessons_dir, prompt_user = TRUE)
}
\arguments{
\item{new_proj_name}{The new name you want to give the selected project}

\item{gh_proj_name}{The unique project title of this lesson as it is named on \url{https://github.com/galacticpolymath}. Not \emph{necessarily} the same as the ShortTitle used in naming lesson presentations and worksheets; probably more specific with underscores. No default.}

\item{lessons_dir}{path to the virtualized folder Edu/lessons, where all the lessons are found; default="/Volumes/GoogleDrive/My Drive/Edu/Lessons"}

\item{prompt_user}{do you want to ask user if they def want to rename this before doing so? default=TRUE}
}
\description{
This is for renaming the GitHub remote repository \strong{after} you've renamed the project folder locally. Uses the gh command line interface (which needs to be set up) and \code{\link[gh:gh]{gh::gh()}} to check existence of the repo and change it online.
}
