% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_reset_remote.R
\name{gh_reset_remote}
\alias{gh_reset_remote}
\title{Reset Remote GitHub Remote Repository}
\usage{
gh_reset_remote(new_name, WD = getwd(), check_current = FALSE)
}
\arguments{
\item{new_name}{The name of the new (empty) repo you want to connect to (should be exactly as it is named on \url{https://github.com/galacticpolymath}); It's just the project name--a full URL is not expected.}

\item{WD}{working directory; default= getwd()}

\item{check_current}{do you want to check whether the current listed GitHubPath in the front-matter.yml is good? default=F}
}
\description{
Change the GitHub URL associated with this project (i.e. after you rename the repo on \url{https://github.com/galacticpolymath}) or after you've "forked" a project to make a new language version of the project.
}
\details{
Will run some validation checks and let you know whether the new_name repo exists on the web. May also prompt to run \code{\link[=editor]{editor()}} to initialize some fields.
}
\seealso{
Other GitHub Functions: 
\code{\link{gh_push_forked_repo}()}
}
\concept{GitHub Functions}
