% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_api_batch_unit_replace.R
\name{gp_api_batch_unit_replace}
\alias{gp_api_batch_unit_replace}
\title{batch replace units}
\usage{
gp_api_batch_unit_replace(shared_drive = "?")
}
\arguments{
\item{shared_drive}{passed to \code{\link[=pick_lesson]{pick_lesson()}} which shared drive do you want to find the lessons in? default= "s" Options:
\itemize{
\item "s" or "?" = GP-Studio (draft working directory, many users with access)
\item "l" or "??" = GP-Live (private, admin only)
}}
}
\description{
Deletes and replaces the selected units
}
