% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_api_get_token.R
\name{gp_api_get_token}
\alias{gp_api_get_token}
\alias{get_gp_api_token}
\title{Get a JWT token for GP API}
\usage{
gp_api_get_token(refresh = FALSE, dev = TRUE)

get_gp_api_token(refresh = FALSE, dev = TRUE)
}
\arguments{
\item{refresh}{do you want to re-authenticate? default=FALSE}

\item{dev}{logical; if TRUE (default), gets catalog from the dev gp-catalog. Otherwise, from the prod catalog.}
}
\value{
invisibly returns the token
}
\description{
Will get a token from the R environment if available and test it with an empty query to GP API. If missing or expired, will attempt to renew through Google Oauth flow.
}
\section{Functions}{
\itemize{
\item \code{get_gp_api_token()}: alias

}}
\seealso{
Other GP API: 
\code{\link{gp_api_query}()},
\code{\link{gp_api_unit_delete}()},
\code{\link{gp_api_unit_insert}()},
\code{\link{gp_api_unit_replace}()}
}
\concept{GP API}
