% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_api_query.R
\name{gp_api_query}
\alias{gp_api_query}
\alias{query_gp_api}
\title{Get info for Galactic Polymath units}
\usage{
gp_api_query(
  keys = c("numID", "_id", "Title"),
  numID = NULL,
  id = NULL,
  output = "tibble"
)

query_gp_api(
  keys = c("numID", "_id", "Title"),
  numID = NULL,
  id = NULL,
  output = "tibble"
)
}
\arguments{
\item{keys}{character vector; which front-matter keys do you want from lessons? default: c("numID","_id","Title"). See all options with \code{\link[=get_fm_names]{get_fm_names()}}}

\item{numID}{is a vector of numIDs for unit(s) you want. default=NULL returns all units}

\item{id}{is a vector of \verb{_id}s for unit(s) you want. default=NULL returns all units}

\item{output}{return values as "tibble" or "list"; default="tibble"}
}
\value{
list of results or tbl_json
}
\description{
Constructs a query and requests selected information from the from the galacticpolymath.com/api
}
\section{Functions}{
\itemize{
\item \code{query_gp_api()}: Alias for gp_api_query

}}
\concept{GP API}
