% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_api_query_users.R
\name{gp_api_query_users}
\alias{gp_api_query_users}
\title{Get GP user profile info}
\usage{
gp_api_query_users(
  WD = "?",
  unit_id = NULL,
  prompt_user = TRUE,
  dev = FALSE,
  verbosity = 1
)
}
\arguments{
\item{prompt_user}{logical; ask user before deleting and replacing the unit? default=TRUE}

\item{dev}{logical; if FALSE (default), gets catalog from the production gp-catalog. Otherwise, from the dev catalog.}

\item{verbosity}{passed to \code{\link[httr2:req_perform]{httr2::req_perform()}}; default=1}
}
\value{
success (logical)
}
\description{
Gets a summary of galacticpolymath.com user accounts through the GP API (needs authenticated user)
}
\seealso{
Other GP API: 
\code{\link{gp_api_get_token}()},
\code{\link{gp_api_query}()},
\code{\link{gp_api_unit_delete}()},
\code{\link{gp_api_unit_insert}()},
\code{\link{gp_api_unit_replace}()}
}
\concept{GP API}
