% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_api_unit_delete.R
\name{gp_api_unit_delete}
\alias{gp_api_unit_delete}
\title{Delete a GP unit}
\usage{
gp_api_unit_delete(WD = "?")
}
\arguments{
\item{WD}{working directory, passed to \code{\link[=parse_wd]{parse_wd()}};default="?"}
}
\value{
success (logical)
}
\description{
Deletes a unit from MongoDB using the GP API
}
