% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_api_unit_delete.R
\name{gp_api_unit_delete}
\alias{gp_api_unit_delete}
\title{Delete a GP unit from the database}
\usage{
gp_api_unit_delete(
  WD = "?",
  unit_id = NULL,
  prompt_user = TRUE,
  dev = FALSE,
  verbosity = 1
)
}
\arguments{
\item{WD}{working directory, passed to \code{\link[=parse_wd]{parse_wd()}};default="?"}

\item{unit_id}{instead of supplying WD, you can supply the _id directly for the unit that you want to delete.}

\item{prompt_user}{logical; ask user before deleting and replacing the unit? default=TRUE}

\item{dev}{logical; if FALSE (default), gets catalog from the production gp-catalog. Otherwise, from the dev catalog.NULL not allowed!}

\item{verbosity}{passed to \code{\link[httr2:req_perform]{httr2::req_perform()}}; default=1}
}
\value{
success (logical)
}
\description{
Deletes a unit from MongoDB using the GP API. Does not delete it from Google Drive or GitHub gp-lessons project.
}
\seealso{
Other GP API: 
\code{\link{gp_api_get_token}()},
\code{\link{gp_api_query}()},
\code{\link{gp_api_unit_insert}()},
\code{\link{gp_api_unit_replace}()}
}
\concept{GP API}
