% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_api_unit_insert.R
\name{gp_api_unit_insert}
\alias{gp_api_unit_insert}
\title{Insert a new GP unit into the database}
\usage{
gp_api_unit_insert(WD = "?")
}
\arguments{
\item{WD}{working directory, passed to \code{\link[=parse_wd]{parse_wd()}};default="?"}
}
\value{
success (logical)
}
\description{
Insert (i.e. add) a new unit to MongoDB using the GP API.
}
\seealso{
Other GP API: 
\code{\link{get_gp_api_token}()},
\code{\link{gp_api_query}()},
\code{\link{gp_api_unit_delete}()},
\code{\link{gp_api_unit_replace}()}
}
\concept{GP API}
