% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_api_unit_replace.R
\name{gp_api_unit_replace}
\alias{gp_api_unit_replace}
\title{Replace a GP unit in the database}
\usage{
gp_api_unit_replace(WD = "?")
}
\arguments{
\item{WD}{working directory, passed to \code{\link[=parse_wd]{parse_wd()}};default="?"}
}
\value{
success (logical)
}
\description{
Completely deletes and re-inserts a LESSON.json for the mini-unit.
}
\details{
Shallow wrapper for internal functions \code{\link[=gp_api_unit_delete]{gp_api_unit_delete()}} and \code{\link[=gp_api_unit_insert]{gp_api_unit_insert()}}
}
\seealso{
Other GP API: 
\code{\link{gp_api_get_token}()},
\code{\link{gp_api_query}()},
\code{\link{gp_api_unit_delete}()},
\code{\link{gp_api_unit_insert}()}
}
\concept{GP API}
