% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_footer.R
\name{gp_footer}
\alias{gp_footer}
\title{gp_footer}
\usage{
gp_footer(
  obj,
  caption,
  x = 0.01,
  y = 0.0275,
  text_size = 7,
  text_col = "#363636",
  data_attrib = NULL,
  data_attrib_scale = 0.6,
  logo = "black",
  fill_col = gpColors("sparkle")
)
}
\arguments{
\item{obj}{ggplot object}

\item{caption}{plot caption}

\item{x}{x coord}

\item{y}{y coord}

\item{text_size}{text size; default=9;}

\item{data_attrib}{concise citation for data source (will be smaller font)}

\item{data_attrib_scale}{default=0.6 from 0 to 1, how much smaller to scale attribution text?}

\item{logo}{default="black"; which GP logo do you want to use?}

\item{fill_col}{fill color for caption at bottom; default is gpColors("sparkle white")}

\item{text_color}{text color; default="#363636" aka Galactic Black}
}
\description{
Add a Galactic Polymath footer to the bottom of ggplot objects
}
