% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpColors.R
\name{gpColors}
\alias{gpColors}
\title{gpColors}
\usage{
gpColors(colorNames, list = FALSE)
}
\arguments{
\item{colorNames}{a vector of quoted named colors (partial name matching is allowed)}

\item{list}{T/F, do you want to list the options?}
}
\value{
a vector of hex codes for the requested colors
}
\description{
output a vector of hex codes, given named colors; list=TRUE lists all options
}
