% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpsave.R
\name{gpsave}
\alias{gpsave}
\title{gpsave}
\usage{
gpsave(filename, width = 7, height = 3.94, dpi = 300, bg = "transparent", ...)
}
\arguments{
\item{filename}{expects filename, with file extension (e.g. "plot.png"); can also include subfolder (e.g. "newfolder/plot.png")}

\item{width}{plot width in inches (default= 7)}

\item{height}{plot height in inches (default= 3.94, i.e. 16:9 aspect ratio)}

\item{dpi}{dots per inch resolution (default= 300); note changing this will unfortunately also change the output relative text size}

\item{bg}{background color (default= "transparent")}

\item{...}{other parameters from \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}
}
\description{
Wrapper for \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} with galacticPubs defaults
}
\details{
Most important thing it does is default to saving in the assets/_R_outputs/ folder and save with reasonable dimensions that maintain large, readable text size of graph labels
}
