% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_fm.R
\name{init_fm}
\alias{init_fm}
\title{Initialize Front Matter}
\usage{
init_fm(WD = "?", WD_git = NULL)
}
\arguments{
\item{WD}{the working directory of the lesson project; default=getwd()}

\item{WD_git}{(alternate; do WD OR WD_git, not both); path to root of project in the gp-lessons/Lessons/Project github repo}
}
\value{
logical of success; T= exists or created; F= not found and not created
}
\description{
Will run \code{\link[=check_fm]{check_fm()}} and if meta/front-matter.yml not found, it will create this file from the galacticPubs template.
}
