% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inSync.R
\name{inSync}
\alias{inSync}
\title{inSync}
\usage{
inSync(path1, path2, verbose = FALSE)
}
\arguments{
\item{path1}{path to file of interest (the one that may or may not need updating)}

\item{path2}{path to reference file (expected to be at least slightly older)}

\item{verbose}{print out table with information? default=FALSE}
}
\value{
T if timestamps match, F if they don't, and NA if one of them doesn't exist
}
\description{
Checks if 2 files exist, and whether path1 is the same age or newer than path2
}
