% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inSync.R
\name{inSync}
\alias{inSync}
\title{inSync}
\usage{
inSync(path1, path2, newer = TRUE, ..., verbose = FALSE, WD = getwd())
}
\arguments{
\item{path1}{path to file of interest (e.g. that is created from the other paths, for newer=T logic)}

\item{path2}{path to reference file (expected to be at least slightly older, if newer=T). Path 2 can also be}

\item{newer}{logical; is path1 expected to be newer than (i.e. derived from) other paths? default=T. If newer=F, will test if path1 is older than all other paths.}

\item{...}{path to other reference files to compare modified date to path1; separated by commas}

\item{verbose}{print out table with information? default=FALSE}

\item{WD}{is working directory of the project (useful to supply for shiny app, which has diff. working environment)}
}
\value{
T if timestamps match, F if they don't or if path 1 is missing
}
\description{
Checks if all files exist, and whether path 1 (the most derived file) is the same age or newer than path 2- path n (which are used to create path 1)
}
