% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\name{is_empty}
\alias{is_empty}
\title{is_empty}
\usage{
is_empty(x, ignore_na_df = FALSE)
}
\arguments{
\item{x}{any value or vector}

\item{ignore_na_df}{logical, if T, when an empty data frame is supplied for \emph{x}, it will return FALSE. Useful if you consider having column information in a tibble not "empty"; default= FALSE}
}
\value{
TRUE (if empty) or FALSE (if not empty)
}
\description{
Logical test for all the different versions of empty values, i.e.:
\itemize{
\item NULL
\item ""
\item NA
\item "\verb{\\\\n}"
\item list()
\item length(x)==0
}
}
\details{
Works on values, vectors, or a flat list. For vector and list inputs output will be a single T or F, not a vector of T/F values.
}
\examples{
is_empty("")
is_empty(3)
is_empty(c("",NA,NULL))
is_empty(list(a="",b=NA,c=NULL))
is_empty(list(a="",b=NA,c="nonempty cell"))
}
