% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\name{is_empty}
\alias{is_empty}
\title{is_empty}
\usage{
is_empty(x, names_meaningful = FALSE)
}
\arguments{
\item{x}{any value or vector}

\item{names_meaningful}{logical, if T, when an empty data frame or list is supplied for \emph{x} that is all NA, but has names, it will return FALSE. default= FALSE}
}
\value{
TRUE (if empty) or FALSE (if not empty)

logical T=Empty, F=Not empty
}
\description{
Logical test for all the different versions of empty values, i.e.:
\itemize{
\item NULL
\item ""
\item NA
\item "\verb{\\\\n}"
\item list()
\item length(x)==0
}
}
\details{
Works on values, vectors, or a flat list. For vector and list inputs output will be a single T or F, not a vector of T/F values.
}
\examples{
is_empty("")
is_empty(3)
is_empty(c("",NA,NULL))
is_empty(list(a="",b=NA,c=NULL))
is_empty(list(a="",b=NA,c="nonempty cell"))

b <- dplyr::tibble(ImportantColumn=NA,ImportantColumn2=NA)
is_empty(b)
#Here, if we care about column names and don't consider this
#empty, we should return FALSE
is_empty(b, names_meaningful=TRUE)

is_empty(list(a=NA,b=b,c=list(),d=NULL))
is_empty(list(a=NA,b=b,c=list(),d=NULL), names_meaningful=TRUE)

#
}
