% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\name{is_empty}
\alias{is_empty}
\title{is_empty}
\usage{
is_empty(x)
}
\arguments{
\item{x}{any value or vector}
}
\value{
TRUE (if empty) or FALSE (if not empty)
}
\description{
Logical test for all the different versions of empty values (NULL, "", NA, "`\\n`",list(),length(x)==0)
}
