% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningChart.R
\name{learningChart}
\alias{learningChart}
\title{learningChart}
\usage{
learningChart(
  caption = NA,
  captionN = TRUE,
  centralText = NA,
  quotedTitle = NA,
  centralTextSize = 3.7,
  saveFile = TRUE,
  destFolder,
  fileName = "GP-Learning-Chart",
  WD = getwd(),
  dpi = 200,
  showPlot = TRUE,
  ...
)
}
\arguments{
\item{caption}{quoted text you want to go at the bottom of the chart}

\item{captionN}{T/F, add the range of the number of standards per grade used to make the plot to caption?}

\item{centralText}{specify grades the chart is for; by default pulls most common gradeBand from compiledAlignment (e.g. "grades\verb{\\\\n}5-6")}

\item{quotedTitle}{the quoted title used to attribute the learning chart (e.g. Knowledge and skills taught by 'quotedTitle')}

\item{centralTextSize}{multiplier for font size of centralText}

\item{saveFile}{T/F, save file or just print to screen?}

\item{destFolder}{where do you want to save the folder; by default in the "assets/learningPlots" folder, 1 level up from the working directory}

\item{fileName}{expects "somefilename" (file extension will be ignored)}

\item{WD}{is working directory of the project (useful to supply for shiny app, which has diff. working environment)}

\item{dpi}{what resolution would you like the output in dots per inch? 300 by default}

\item{showPlot}{plot to screen or just save to file? default=T}

\item{...}{other arguments passed to \code{\link[grDevices]{png}}}
}
\value{
the learning chart plot object (grid format); Result can be plotted with grid::grid.draw(result). The chart file is saved to assets/GP_Learning_Chart.png by default
}
\description{
Make a GP Learning Chart
}
