% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningChart.R
\name{learningChart}
\alias{learningChart}
\title{learningChart}
\usage{
learningChart(
  compiledAlignment,
  targetSubj = NULL,
  caption,
  captionN = T,
  centralText = "grades\\n5-12",
  centralTextSize = 3.7,
  saveFile = TRUE,
  destFolder = "assets/learning-plots/",
  fileName = "GP-Learning-Chart",
  dpi = 300,
  ...
)
}
\arguments{
\item{compiledAlignment}{the output of \code{\link{compileStandards}}}

\item{targetSubj}{which subject(s) is (are) the focus of the lesson? opts= "math","ela","science","social studies"}

\item{caption}{quoted text you want to go at the bottom of the chart}

\item{captionN}{T/F, add the range of the number of standards per grade used to make the plot to caption?}

\item{centralText}{what you want at the center of the plot ("grades 5-12" by default)}

\item{centralTextSize}{multiplier for font size of centralText}

\item{saveFile}{T/F, save file or just print to screen?}

\item{destFolder}{where do you want to save the folder; by default in the "assets/learningPlots" folder, 1 level up from the working directory}

\item{fileName}{expects "somefilename" (file extension will be ignored)}

\item{dpi}{what resolution would you like the output in dots per inch? 300 by default}

\item{...}{other arguments passed to \code{\link[grDevices]{png}}}
}
\value{
the learning chart plot object (grid format); the file is saved to assets/GP_Learning_Chart.png by default
}
\description{
Make a GP Learning Chart
}
