% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningEpaulette.R
\name{learningEpaulette}
\alias{learningEpaulette}
\title{learningEpaulette}
\usage{
learningEpaulette(
  WD = "?",
  heightScalar = 0.5,
  epauletteHeight = 0.2,
  randomSeed = 101,
  saveFile = TRUE,
  destFolder,
  fileName = "GP-Learning-Epaulette",
  font_size = 10,
  width = 11,
  height = 1.6,
  dpi = 200,
  showPlot = TRUE,
  ...
)
}
\arguments{
\item{WD}{working directory; default=getwd(); if "?" supplied, will invoke \code{\link[=pick_lesson]{pick_lesson()}}. The basename of this working directory will then be used to find a match in the gp-lessons git project folder by calling \code{\link[=get_wd_git]{get_wd_git()}}. It's a little roundabout, but is consistent with lookups centering on the Google Drive project working directory.}

\item{heightScalar}{for horizontal epaulette, multiplier for image height which affects amount of padding between label text and epaulette; default=1}

\item{epauletteHeight}{relative size of the epaulette; default=0.2}

\item{randomSeed}{random number for getting slightly different (but repeatable) repelled text labels}

\item{saveFile}{T/F, save file or just print to screen?}

\item{destFolder}{where do you want to save the folder; by default in the "assets/_learning-plots" folder, 1 level up from the working directory}

\item{fileName}{expects "somefilename" for ggsave output image file}

\item{font_size}{size of font in pts;default=11}

\item{width}{plot width in inches}

\item{height}{plot height in inches}

\item{dpi}{resolution in dots per inch; by default 150}

\item{showPlot}{plot to screen or just save to file? default=T}

\item{...}{additional parameters for \code{\link[ggplot2]{ggsave}}}
}
\value{
returned plot as a ggplot object; plot saved to assets/GP_Learning_Epaulette.png by default
}
\description{
Create a Galactic Polymath Learning Epaulette which is a special kind of mosaic plot showing proportion of lesson by subject. Outputs a horizontal and a vertical version for mobile
}
