% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningEpaulette.R
\name{learningEpaulette}
\alias{learningEpaulette}
\title{learningEpaulette}
\usage{
learningEpaulette(
  compiledAlignment,
  targetSubj = NULL,
  vertSpacing = c(1, 1, 1, 1),
  saveFile = TRUE,
  destFolder = "assets/learning-plots/",
  fileName = "GP-Learning-Epaulette",
  thickness = 0.2,
  width = 11,
  height = 1.6,
  dpi = 200,
  ...
)
}
\arguments{
\item{compiledAlignment}{the output of \code{\link{compileAlignment}}}

\item{targetSubj}{which subject`(`s`)` is `(`are`)` the focus of the lesson? opts= "math","ela","science","social studies"}

\item{vertSpacing}{4 value vector ranging from 0 to 1 for manipulating label spacing}

\item{saveFile}{T/F, save file or just print to screen?}

\item{destFolder}{where do you want to save the folder; by default in the "assets/learningPlots" folder, 1 level up from the working directory}

\item{fileName}{expects "somefilename" for ggsave output image file}

\item{thickness}{is how thick to make the epaulette bar `(`range from 0 to 0.5`)`, 0.2 or 20\% of vertical plot space by default}

\item{width}{plot width in inches}

\item{height}{plot height in inches}

\item{dpi}{resolution in dots per inch; by default 150}

\item{...}{additional parameters for \code{\link[ggplot2]{ggsave}}}
}
\value{
returned plot as a ggplot object; plot saved to assets/GP_Learning_Epaulette.png by default
}
\description{
Create a Galactic Polymath Learning Epaulette which is a special kind of mosaic plot showing proportion of lesson by subject
}
