% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesson_new_locale.R
\name{lesson_new_locale}
\alias{lesson_new_locale}
\title{Make a new locale version of existing lesson}
\usage{
lesson_new_locale(gh_proj_name)
}
\arguments{
\item{gh_proj_name}{the name of the project as it is on GitHub and Google Drive}
}
\value{
summary tibble
}
\description{
This function will:
\enumerate{
\item ask user to specify locale
\item clone a lesson found on GitHub
\item fill in missing files in the Google Drive folder
}
}
