% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lessons_get_path.R
\name{lessons_get_path}
\alias{lessons_get_path}
\alias{get_lessons_path}
\title{Get path to Edu/Lessons folder containing GP projects}
\usage{
lessons_get_path(shared_drive = "s")

get_lessons_path(shared_drive = "s")
}
\arguments{
\item{shared_drive}{which shared drive do you want to find the lessons in? default= "s" Options:
\itemize{
\item "s" or "?" = GP-Studio (draft working directory, many users with access)
\item "l" or "??" = GP-Live (private, admin only)
\item "gp"= GalacticPolymath (public-facing read-only)
\item "git" = wd_git (i.e. the github repository for the gp-lessons project)
}}
}
\description{
Retrieves an environmental variable set with \code{\link[=init_galacticPubs]{init_galacticPubs()}}
}
\section{Functions}{
\itemize{
\item \code{get_lessons_path()}: Alias for \code{\link[=lessons_get_path]{lessons_get_path()}}

}}
