% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shareable_assets.R
\name{make_shareable_assets}
\alias{make_shareable_assets}
\title{make_shareable_assets}
\usage{
make_shareable_assets(WD = "?")
}
\arguments{
\item{WD}{"?"}
}
\value{
success
}
\description{
Make an HTML file with shareable assets
}
