% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_yt_embed.R
\name{make_yt_embed}
\alias{make_yt_embed}
\title{make_yt_embed}
\usage{
make_yt_embed(link)
}
\arguments{
\item{link}{a YouTube link}
}
\description{
Internal helpers for \code{\link[=compile_teach_it]{compile_teach_it()}}. Get YT embed code from any Youtube link
}
\examples{
#Should give 3 URLs in format www.youtube.com/embed/videoID and 1 original URL (not on YT)
c("https://youtu.be/mD24yi7uLgU","https://youtu.be/ZAFjYJk27Ug?si=d-MIziNb39ib_8fW",
"https://www.youtube.com/watch?v=h5eTqjzQZDY",
"https://into-the-dark.galacticpolymath.com/") \%>\% make_yt_embed()
#should not change anything
make_yt_embed("https://www.youtube.com/embed/A2FBO8ofEac")

}
\seealso{
Other Internal helper functions: 
\code{\link{zcatchLinkNA}()},
\code{\link{zget_envir}()}
}
\concept{Internal helper functions}
