% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_proj_folder.R
\name{open_proj_folder}
\alias{open_proj_folder}
\title{Open project local (virtualized) project folder}
\usage{
open_proj_folder(subpath = NULL, WD = "?")
}
\arguments{
\item{subpath}{subpath to concat to WD path e.g.c("assets","data","teaching-materials"); default=NULL}

\item{WD}{working directory; passed to \code{\link[=parse_wd]{parse_wd()}}.}
}
\description{
Requires Google Drive for Desktop to be set up and to have access to the folder. If it's on GP-LIVE, you may not have access.
}
