% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_wd.R
\name{parse_wd}
\alias{parse_wd}
\title{parse_wd}
\usage{
parse_wd(str = NULL, use_wd_git = FALSE, ...)
}
\arguments{
\item{str}{is a string supplied as the WD parameter in a galacticPubs function. Often "?"}

\item{use_wd_git}{passed to \code{\link[=pick_lesson]{pick_lesson()}}; default=FALSE}

\item{...}{other parameters passed to \code{\link[=pick_lesson]{pick_lesson()}}}
}
\value{
if str is not "?", "??","???",or "?!", it will return str. Otherwise, returns the result of \code{\link[=pick_lesson]{pick_lesson()}}
}
\description{
Helper function to handle Working Directory strings. Here's what happens when you supply the following:
\itemize{
\item "?" or "?s": invokes \verb{[pick_lesson("s")]} to choose from among lessons in the 'GP-Studio' shared drive.
\item "??" or "?l": invokes \verb{[pick_lesson("l")]} to pick a path to lessons on the 'GP-LIVE' drive.
\item "???" or "?gp": gives you \verb{[pick_lesson("gp")]} to pick a path to lesson teaching materials found on 'GalacticPolymath' drive
\item "?!" or "sl": invokes \verb{[pick_lesson("sl")]} to pick from units on 'GP-Studio' and 'GP-LIVE'
}
}
