% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGPmarkdown.R
\name{parseGPmarkdown}
\alias{parseGPmarkdown}
\title{parseGPmarkdown}
\usage{
parseGPmarkdown(x, linksFile = "meta/teaching-material-links.xlsx")
}
\arguments{
\item{x}{a text string to parse}

\item{linksFile}{file location of the lesson teaching-material-links XLSX worksheet. This is used for our custom
markdown; e.g. "\{vid1\}" will be replaced with a markdown link to the video in the links spreadsheet multimedia tab
that has order=1}
}
\value{
formatted replacement text
}
\description{
Function to replace custom expressions with appropriate links or text
}
