% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGPmarkdown.R
\encoding{UTF-8}
\name{parseGPmarkdown}
\alias{parseGPmarkdown}
\title{parseGPmarkdown}
\usage{
parseGPmarkdown(x, linksFile = "meta/teaching-materials.xlsx", WD = NULL)
}
\arguments{
\item{x}{a text string to parse}

\item{linksFile}{file location of the lesson teaching-materials XLSX worksheet. This is used for our custom
markdown; e.g. "\{vid1\}" will be replaced with a markdown link to the video in the links spreadsheet multimedia tab
that has order=1}

\item{WD}{working directory; default=NULL}
}
\value{
formatted replacement text
}
\description{
Functions to replace custom expressions with appropriate links or text. Custom tags include "\{vid 1\}", "\{vid 2\}" or "\{item 1\}", "\{item 2\}" more generally. Both have the same effect, but vid is clearer when reading and enforces video type, while item is more flexible. Symbols for media types: 'video'= ▶, 'pdf' or default= ➚. Special case for media titles with keyword "Cards"= ♧
}
\seealso{
Other markdown: 
\code{\link{expand_md_links}()},
\code{\link{fixAnchorLinks}()}
}
\concept{markdown}
