% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_valid.R
\name{paste_valid}
\alias{paste_valid}
\title{paste_valid}
\usage{
paste_valid(..., sep = "", collapse = " ")
}
\arguments{
\item{...}{a set of arguments to be combined into a single string}

\item{sep}{character added to end of each argument; default sep=""}

\item{collapse}{character separating the concatenated strings; default collapse=" "}
}
\description{
Given a set of character objects, combine all non-empty variables into a single string. Uses \code{\link[=is_empty]{is_empty()}} to test validity, so should be robust to all kinds of missingness. It will remove both NA and NULL.
}
\seealso{
Other simple text string manipulation functions: 
\code{\link{string_capitalize_first}()},
\code{\link{string_parseCamel}()}
}
\concept{simple text string manipulation functions}
