% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_parent_dir.R
\name{path_parent_dir}
\alias{path_parent_dir}
\title{Find the parent directory for a path}
\usage{
path_parent_dir(filepath)
}
\arguments{
\item{filepath}{the file path}
}
\description{
Supply a path, get the parent folder
}
