% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_lesson.R
\name{pick_lesson}
\alias{pick_lesson}
\title{pick_lesson}
\usage{
pick_lesson(lessons_dir)
}
\arguments{
\item{lessons_dir}{the path to the directory where lessons are held (make sure it leads with a /)}
}
\value{
the selected lesson name
}
\description{
Gives a list of lesson options for you to choose from, given the EDU/lessons path
}
