% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_lesson.R
\name{pick_lesson}
\alias{pick_lesson}
\alias{lesson_pick}
\title{pick_lesson (or lesson_pick)}
\usage{
pick_lesson(
  shared_drive = "s",
  show_all = FALSE,
  pick_all = FALSE,
  full_path = TRUE,
  sort_col = "LastUpdated",
  sort_decr = TRUE,
  lessons_dir = NULL,
  pull_path = TRUE,
  use_wd_git = FALSE,
  exclude_TEST = TRUE
)

lesson_pick(
  shared_drive = "s",
  show_all = FALSE,
  pick_all = FALSE,
  full_path = TRUE,
  sort_col = "LastUpdated",
  sort_decr = TRUE,
  lessons_dir = NULL,
  pull_path = TRUE,
  use_wd_git = FALSE,
  exclude_TEST = TRUE
)
}
\arguments{
\item{shared_drive}{which shared drive do you want to find the lessons in? default= "s" Options:
\itemize{
\item "s" or "?" = GP-Studio (draft working directory, many users with access)
\item "l" or "??" = GP-Live (private, admin only)
\item "gp"= GalacticPolymath (public-facing read-only)
\item "sl" or "?!" = both GP-Studio and GP-Live
}}

\item{show_all}{show an 'all' option? default=FALSE}

\item{pick_all}{logical; Default=FALSE makes user pick unit. If TRUE, choice is set to "all" and will return paths to all projects in this directory.}

\item{full_path}{do you want a full path to the chosen lesson? default= TRUE}

\item{sort_col}{which column of output to sort by? Default "LastUpdated"; options=c("Unit","numID","ReleaseDate")}

\item{sort_decr}{logical; sort by decreasing order? default= TRUE}

\item{lessons_dir}{the path to the directory where lessons are held (make sure it leads with a /); default=NULL will resolve by calling \code{\link[=lessons_get_path]{lessons_get_path()}}}

\item{pull_path}{logical; do you want to pull (and unlist) only the paths? default= TRUE. FALSE will return the tibble for the selected units}

\item{use_wd_git}{if TRUE, will use the git root of the current working directory as the base path for relative paths. Makes no expectation that front-matter.yml is present or working (useful in combination with \code{\link[=fm_restore]{fm_restore()}} when YML is corrupted)}

\item{exclude_TEST}{default=T; excludes test repositories}
}
\value{
the selected path, tibble, or string of names
}
\description{
Interactively lets you pick from a list of published Galactic Polymath units and will out put a virtualized Google Drive for Desktop path
}
\section{Functions}{
\itemize{
\item \code{lesson_pick()}: 

}}
