% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_user_stats.R
\name{report_user_stats}
\alias{report_user_stats}
\title{Report information about GP users}
\usage{
report_user_stats()
}
\value{
Summary tibble
}
\description{
Provide basic information about users
}
