% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_user_stats.R
\name{report_user_stats}
\alias{report_user_stats}
\title{Report information about GP users}
\usage{
report_user_stats(verbosity = 1, view_data = TRUE)
}
\arguments{
\item{verbosity}{passed to \code{\link[httr2:req_perform]{httr2::req_perform()}}; default=1}

\item{view_data}{logical; do you want to open retrieved user data with \code{\link[=View]{View()}}? Default=TRUE}
}
\value{
Summary tibble
}
\description{
Provide basic information about users
}
