% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_user_stats.R
\name{report_user_stats}
\alias{report_user_stats}
\title{Report information about GP users}
\usage{
report_user_stats(verbosity = 1)
}
\arguments{
\item{verbosity}{passed to \code{\link[httr2:req_perform]{httr2::req_perform()}}; default=1}
}
\value{
Summary tibble
}
\description{
Provide basic information about users
}
