% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLessonScripts.R
\name{runLessonScripts}
\alias{runLessonScripts}
\title{runLessonScripts}
\usage{
runLessonScripts(skip = NULL)
}
\arguments{
\item{skip}{filename of a script to skip}
}
\description{
Run all scripts in the scripts/ subfolder in the lesson directory. Scripts must not have any interdependencies, as a general design rule.
}
