% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_read_yaml.R
\name{safe_read_yaml}
\alias{safe_read_yaml}
\title{safe_read_yaml}
\usage{
safe_read_yaml(yaml_path, eval.expr = TRUE)
}
\arguments{
\item{yaml_path}{path to the front-matter.yml file (usually in working_directory/meta/)}

\item{eval.expr}{boolean; do you want to evaluate expression in yaml prepended with '!expr '? Default=TRUE}
}
\description{
Wrapper for \link[yaml:read_yaml]{read_yaml} that simplifies all null and missing data to ""
}
