% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_assets.R
\name{stage_assets}
\alias{stage_assets}
\title{Copy updated assets in a specific folder to a destination}
\usage{
stage_assets(
  WD,
  rel_path,
  pattern = NULL,
  exclude = NULL,
  dest_folder = NULL,
  clear = FALSE
)
}
\arguments{
\item{WD}{what's the project working directory? passed to parse_wd()}

\item{rel_path}{what's the subfolder path within WD? e.g."assets/_learning-plots"}

\item{pattern}{a regular expression for allowable file types; default=NULL}

\item{exclude}{a regular expression to exclude from uploading; default=NULL excludes nothing; ".ai$" would exclude Adobe Illustrator files for example.}

\item{dest_folder}{full path to where you want things to go (defaults to WD/published)}

\item{clear}{do you want to delete everything in the target directory? default= FALSE}
}
\description{
Helper for \code{\link[=upload_assets]{upload_assets()}}. Not really intended for use by user. Looks for a pattern in a given path within a working directory and copies those files to the dest_folder, usually /published
}
