% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ISO_codes.R
\name{update_ISO_codes}
\alias{update_ISO_codes}
\title{update_ISO_codes}
\usage{
update_ISO_codes()
}
\description{
Internal function for downloading the latest ISO codes using the ISOcodes package. Run with `galacticPubs:::update_ISO_codes()`
}
