% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTeachingMatLinks.R
\name{updateTeachingMatLinks}
\alias{updateTeachingMatLinks}
\title{updateTeachingMatLinks}
\usage{
updateTeachingMatLinks(
  shortTitle,
  dataCat = c("download", "remote", "classroom"),
  linksFile = "teaching-materials.xlsx",
  WD = getwd(),
  sortOutput = T,
  returnWorkbook = F
)
}
\arguments{
\item{shortTitle}{The unique short title of this lesson which is prefixed on the lesson folder name in the shared}

\item{dataCat}{which info do you want to merge with your teaching-materials spreadsheet? Options= "download", "classroom" and "remote". Default is all. Abbreviation with first letters acceptable.}

\item{linksFile}{name of the file we're updating in the meta/ subfolder; default="teaching-materials.xlsx". gdrive *CaseSensitive!}

\item{WD}{is working directory of the project (useful to supply for shiny app, which has diff. working environment)}

\item{sortOutput}{logical; if T, outputs are sorted by grade, part, filetype, then filename}

\item{returnWorkbook}{Logical; if T, returns the list which was written to linksFile}
}
\description{
Adds/updates Google Drive share links to the requested tabs in teaching-materials.xlsx. If a file is found on Google Drive, it will add it to the appropriate tab, with a timestamp in updateNotes. If a listed google drive link doesn't match a found, an updateNote will be added "trashed?"
}
\details{
Just a note that I'm using 2 libraries to handle XLSX files. Not ideal, but the imported data from openxlsx::read.xlsx is nicer than XLConnect::readWorksheetFromFile, but the editing of data from a complex spreadsheet is MUCH better in XLConnect.
}
