% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_assets.R
\name{upload_assets}
\alias{upload_assets}
\title{upload_assets()}
\usage{
upload_assets(WD = "?", bucket = "gp-cloud")
}
\arguments{
\item{WD}{WD working directory, passed to \code{\link[=parse_wd]{parse_wd()}}}
}
\value{
a Tibble with success, filenames and download links
}
\description{
Upload lesson assets to galacticPubs Google Cloud Storage bucket
}
