% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_assets.R
\name{upload_assets}
\alias{upload_assets}
\alias{gcs_upload_assets}
\title{upload_assets()}
\usage{
upload_assets(WD = "?", bucket = "gp-cloud", clear = FALSE)

gcs_upload_assets(WD = "?", bucket = "gp-cloud", clear = FALSE)
}
\arguments{
\item{WD}{WD working directory, passed to \code{\link[=parse_wd]{parse_wd()}}}

\item{bucket}{which bucket? default= "gp-cloud"}

\item{clear}{do you want to clear all existing files for this project before uploading? default=FALSE}
}
\value{
invisibly returns a Tibble with filenames and download links
}
\description{
Upload lesson assets to galacticPubs Google Cloud Storage bucket

Alias for upload_assets
}
\section{Functions}{
\itemize{
\item \code{gcs_upload_assets()}: alias

}}
\seealso{
Other google cloud storage: 
\code{\link{gcs_add}()},
\code{\link{gcs_contents}()},
\code{\link{gcs_delete}()}

Other google cloud storage: 
\code{\link{gcs_add}()},
\code{\link{gcs_contents}()},
\code{\link{gcs_delete}()}
}
\concept{google cloud storage}
