% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichRepo.R
\name{whichRepo}
\alias{whichRepo}
\title{whichRepo}
\usage{
whichRepo(fullPath = FALSE)
}
\arguments{
\item{fullPath}{do you want to export the full path, or just the repo name? Default=F (just name)}
}
\value{
Either a warning or the name of the github repository connected to the current Rstudio project (extracted from `rstudioapi::getActiveProject()` and a call to `git remote -v`)
}
\description{
Get the name of the repo this is set up on. No error catching at the moment.
}
