% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zget_envir.R
\name{zcatchLinkNA}
\alias{zcatchLinkNA}
\title{zcatchLinkNA}
\usage{
zcatchLinkNA(linkText, url)
}
\arguments{
\item{linkText}{text for the link (describing what it connects to)}

\item{url}{the link URL}
}
\description{
Internal helpers for \code{\link[=compile_teach_it]{compile_teach_it()}}. Change linktext if URL is empty.
}
\seealso{
Other Internal helper functions: 
\code{\link{make_yt_embed}()},
\code{\link{zget_envir}()}
}
\concept{Internal helper functions}
