% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zget_rsrcs.R
\name{zcatchLinkNA}
\alias{zcatchLinkNA}
\title{zcatchLinkNA}
\usage{
zcatchLinkNA(linkText, url)
}
\description{
Internal helpers for \code{\link[=compile_teach_mat]{compile_teach_mat()}}. Change linktext if URL is empty.
}
\seealso{
Other Internal helper functions: 
\code{\link{zYTembed}()},
\code{\link{zget_grade_bands}()},
\code{\link{zget_rsrcs}()}
}
\concept{Internal helper functions}
