% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zget_envir.R
\name{zget_envir}
\alias{zget_envir}
\alias{zget_grade_bands}
\alias{zget_lessons}
\alias{zget_items}
\title{zget_envir()}
\usage{
zget_envir(tlinks, proc_data, lesson_statuses, uinfo, fm)

zget_grade_bands(df_i, proc_data, lesson_statuses, fm, uinfo, assess)

zget_lessons(df_i, proc_data, lesson_statuses, uinfo, fm)

zget_items(df, fm)
}
\arguments{
\item{tlinks}{a tibble of the tlinks of imported teach-it.gsheet}

\item{proc_data}{proc_data}

\item{lesson_statuses}{lesson_statuses}

\item{uinfo}{unit info}

\item{fm}{front matter info from \code{\link[=get_fm]{get_fm()}}}

\item{assess}{tibble of assessment info passed from \code{\link[=zget_envir]{zget_envir()}}}
}
\description{
Internal helpers for \code{\link[=compile_teach_it]{compile_teach_it()}}. Used to pull out highly structured hierarchical JSON from the teach-it.gsheet

Helper function for \code{\link[=compile_teach_it]{compile_teach_it()}} that extracts different kinds of share links for the webpage. For example, makes preview links for google presentations and PDF download links for other documents
}
\details{
Not intended for use outside functions.
}
\section{Functions}{
\itemize{
\item \code{zget_grade_bands()}: 

\item \code{zget_lessons()}: 

\item \code{zget_items()}: 

}}
\seealso{
Other Internal helper functions: 
\code{\link{make_yt_embed}()},
\code{\link{zcatchLinkNA}()}

Other Internal helper functions: 
\code{\link{make_yt_embed}()},
\code{\link{zcatchLinkNA}()}

Other Internal helper functions: 
\code{\link{make_yt_embed}()},
\code{\link{zcatchLinkNA}()}

Other Internal helper functions: 
\code{\link{make_yt_embed}()},
\code{\link{zcatchLinkNA}()}
}
\concept{Internal helper functions}
