% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zget_envir.R
\name{zget_envir}
\alias{zget_envir}
\alias{zget_grade_var_notes}
\alias{zget_grade_bands}
\alias{zget_parts}
\alias{zget_items}
\title{zget_envir()}
\usage{
zget_envir(df, fm)

zget_grade_var_notes(df)

zget_grade_bands(df, fm, assess)

zget_parts(df, fm)

zget_items(df, fm)
}
\arguments{
\item{df}{a dataframe (tibble)}

\item{fm}{front matter info from \code{\link[=get_fm]{get_fm()}}}

\item{assess}{tibble of assessment info passed from \code{\link[=zget_envir]{zget_envir()}}}
}
\description{
Internal helpers for \code{\link[=compile_teach_it]{compile_teach_it()}}. Used to pull out highly structured hierarchical JSON from the teach-it.gsheet

Helper function for \code{\link[=compile_teach_it]{compile_teach_it()}} that extracts different kinds of share links for the webpage. For example, makes preview links for google presentations and PDF download links for other documents
}
\details{
Not intended for use outside functions.
}
\section{Functions}{
\itemize{
\item \code{zget_grade_var_notes()}: 

\item \code{zget_grade_bands()}: 

\item \code{zget_parts()}: 

\item \code{zget_items()}: 

}}
\seealso{
Other Internal helper functions: 
\code{\link{zYTembed}()},
\code{\link{zcatchLinkNA}()}

Other Internal helper functions: 
\code{\link{zYTembed}()},
\code{\link{zcatchLinkNA}()}

Other Internal helper functions: 
\code{\link{zYTembed}()},
\code{\link{zcatchLinkNA}()}

Other Internal helper functions: 
\code{\link{zYTembed}()},
\code{\link{zcatchLinkNA}()}

Other Internal helper functions: 
\code{\link{zYTembed}()},
\code{\link{zcatchLinkNA}()}
}
\concept{Internal helper functions}
