% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zget_procedure.R
\name{zget_procedure}
\alias{zget_procedure}
\title{zget_procedure}
\usage{
zget_procedure(proc, pext, pinfo, mlinks)
}
\arguments{
\item{proc}{the procedure file that has already been read in from the teach-it.gsheet, i.e. via \code{\link[=compile_teach_it]{compile_teach_it()}}}

\item{pext}{the extension info read from the teach-it.gsheet}

\item{pinfo}{the part info read in from the teach-it.gsheet}

\item{mlinks}{multimedia info read in from the teach-it.gsheet}
}
\value{
a list of: the procedure chunks and Part-Extension links for each part and a compiled vocab dataframe
}
\description{
A helper function for \code{\link[=compile_teach_it]{compile_teach_it()}}. Translate the 'Procedure' tab of the lesson's 'teach-it.gsheet' into JSON format.
}
