% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zget_rsrcs.R
\name{zYTembed}
\alias{zYTembed}
\title{zYTembed}
\usage{
zYTembed(link)
}
\arguments{
\item{link}{a YouTube link}
}
\description{
Internal helpers for \code{\link[=compile_teach_mat]{compile_teach_mat()}}. Get YT embed code from any Youtube link
}
\seealso{
Other Internal helper functions: 
\code{\link{zcatchLinkNA}()},
\code{\link{zget_grade_bands}()},
\code{\link{zget_rsrcs}()}
}
\concept{Internal helper functions}
