% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zget_envir.R
\name{zYTembed}
\alias{zYTembed}
\title{zYTembed}
\usage{
zYTembed(link)
}
\arguments{
\item{link}{a YouTube link}
}
\description{
Internal helpers for \code{\link[=compile_teach_it]{compile_teach_it()}}. Get YT embed code from any Youtube link
}
\seealso{
Other Internal helper functions: 
\code{\link{zcatchLinkNA}()},
\code{\link{zget_envir}()}
}
\concept{Internal helper functions}
